Adjust spectral intensities to absorbance units

Converts reflectance or transmittance intensity units to absorbance units.

Usage

adj_intens(x, ...)

# S3 method for formula
adj_intens(formula, data = NULL, ...)

# S3 method for data.frame
adj_intens(x, ...)

# S3 method for default
adj_intens(x, y, type = "none", make_rel = TRUE, ...)

Arguments

x

a numeric vector containing the spectral wavenumbers; alternatively a data frame containing spectral data as "wavenumber" and "intensity" can be supplied.

formula

an object of class 'formula' of the form intensity ~ wavenumber.

data

a data frame containing the variables in formula.

y

a numeric vector containing the spectral intensities.

type

a character string specifying whether the input spectrum is in absorbance units ("none", default) or needs additional conversion from "reflectance" or "transmittance" data.

make_rel

logical; if TRUE spectra are automatically normalized with make_rel().

...

further arguments passed to the submethods.

Value

adj_intens() returns a data frame containing two columns named "wavenumber" and "intensity".

Details

Many of the Open Specy functions will assume that the spectrum is in absorbance units. For example, see match_spec() and subtr_bg(). To run those functions properly, you will need to first convert any spectra from transmittance or reflectance to absorbance using this function. The transmittance adjustment uses the \(log10(1 / T)\) calculation which does not correct for system and particle characteristics. The reflectance adjustment uses the Kubelka-Munk equation \((1 - R)^2 / 2R\). We assume that the reflectance intensity is a percent from 1-100 and first correct the intensity by dividing by 100 so that it fits the form expected by the equation.

See also

Author

Win Cowger, Zacharias Steinmetz

Examples

data("raman_hdpe")

adj_intens(raman_hdpe)
#>     wavenumber  intensity
#> 1      301.040 0.00000000
#> 2      304.632 0.03037975
#> 3      308.221 0.02784810
#> 4      311.810 0.02405063
#> 5      315.398 0.02531646
#> 6      318.983 0.02025316
#> 7      322.566 0.02405063
#> 8      326.150 0.02278481
#> 9      329.732 0.02784810
#> 10     333.311 0.02531646
#> 11     336.889 0.02784810
#> 12     340.467 0.02278481
#> 13     344.042 0.02911392
#> 14     347.618 0.03291139
#> 15     351.190 0.02784810
#> 16     354.760 0.03417722
#> 17     358.332 0.03291139
#> 18     361.900 0.02531646
#> 19     365.466 0.01898734
#> 20     369.031 0.03037975
#> 21     372.597 0.04556962
#> 22     376.160 0.03924051
#> 23     379.721 0.03291139
#> 24     383.280 0.02405063
#> 25     386.839 0.03670886
#> 26     390.396 0.02025316
#> 27     393.951 0.03291139
#> 28     397.505 0.02278481
#> 29     401.058 0.04556962
#> 30     404.610 0.03924051
#> 31     408.160 0.03037975
#> 32     411.708 0.04177215
#> 33     415.256 0.02784810
#> 34     418.801 0.02658228
#> 35     422.346 0.04177215
#> 36     425.887 0.04050633
#> 37     429.429 0.04050633
#> 38     432.970 0.06075949
#> 39     436.507 0.05443038
#> 40     440.044 0.06582278
#> 41     443.581 0.05189873
#> 42     447.116 0.05949367
#> 43     450.647 0.04177215
#> 44     454.180 0.05822785
#> 45     457.708 0.04936709
#> 46     461.238 0.03417722
#> 47     464.764 0.04936709
#> 48     468.291 0.04683544
#> 49     471.815 0.04430380
#> 50     475.339 0.03797468
#> 51     478.859 0.02658228
#> 52     482.379 0.03291139
#> 53     485.899 0.03417722
#> 54     489.416 0.04050633
#> 55     492.931 0.04050633
#> 56     496.447 0.02151899
#> 57     499.960 0.03544304
#> 58     503.471 0.03670886
#> 59     506.981 0.03037975
#> 60     510.489 0.02658228
#> 61     513.999 0.03670886
#> 62     517.503 0.04050633
#> 63     521.009 0.03037975
#> 64     524.510 0.02658228
#> 65     528.014 0.02278481
#> 66     531.513 0.03670886
#> 67     535.014 0.03797468
#> 68     538.510 0.04050633
#> 69     542.008 0.03544304
#> 70     545.502 0.03291139
#> 71     548.997 0.02405063
#> 72     552.488 0.04810127
#> 73     555.979 0.04556962
#> 74     559.469 0.04556962
#> 75     562.955 0.03924051
#> 76     566.442 0.03670886
#> 77     569.928 0.02911392
#> 78     573.411 0.04936709
#> 79     576.894 0.04556962
#> 80     580.376 0.03417722
#> 81     583.854 0.03670886
#> 82     587.331 0.04177215
#> 83     590.811 0.05949367
#> 84     594.286 0.04683544
#> 85     597.759 0.04683544
#> 86     601.231 0.05189873
#> 87     604.705 0.06455696
#> 88     608.174 0.06075949
#> 89     611.642 0.06075949
#> 90     615.109 0.06708861
#> 91     618.574 0.05063291
#> 92     622.038 0.05063291
#> 93     625.503 0.05189873
#> 94     628.966 0.04303797
#> 95     632.426 0.04177215
#> 96     635.884 0.02784810
#> 97     639.342 0.02278481
#> 98     642.798 0.03924051
#> 99     646.252 0.04430380
#> 100    649.706 0.04177215
#> 101    653.158 0.02784810
#> 102    656.608 0.04177215
#> 103    660.058 0.04177215
#> 104    663.506 0.02151899
#> 105    666.953 0.03037975
#> 106    670.398 0.02911392
#> 107    673.842 0.04430380
#> 108    677.285 0.04050633
#> 109    680.727 0.04810127
#> 110    684.165 0.06075949
#> 111    687.602 0.02911392
#> 112    691.040 0.04810127
#> 113    694.476 0.03670886
#> 114    697.911 0.03670886
#> 115    701.345 0.04556962
#> 116    704.775 0.03291139
#> 117    708.204 0.04556962
#> 118    711.634 0.03670886
#> 119    715.063 0.04936709
#> 120    718.490 0.03544304
#> 121    721.912 0.03291139
#> 122    725.336 0.05063291
#> 123    728.760 0.03544304
#> 124    732.180 0.03544304
#> 125    735.599 0.03544304
#> 126    739.018 0.04303797
#> 127    742.434 0.03670886
#> 128    745.849 0.03797468
#> 129    749.265 0.03417722
#> 130    752.675 0.04430380
#> 131    756.088 0.04683544
#> 132    759.498 0.03417722
#> 133    762.906 0.03037975
#> 134    766.313 0.04556962
#> 135    769.719 0.04556962
#> 136    773.124 0.05569620
#> 137    776.527 0.04683544
#> 138    779.929 0.04430380
#> 139    783.330 0.03670886
#> 140    786.729 0.03797468
#> 141    790.127 0.05569620
#> 142    793.524 0.04556962
#> 143    796.919 0.05063291
#> 144    800.312 0.04303797
#> 145    803.705 0.03797468
#> 146    807.097 0.05189873
#> 147    810.485 0.03924051
#> 148    813.874 0.04810127
#> 149    817.262 0.03797468
#> 150    820.647 0.03164557
#> 151    824.032 0.03797468
#> 152    827.416 0.04177215
#> 153    830.797 0.04430380
#> 154    834.179 0.03417722
#> 155    837.557 0.05316456
#> 156    840.936 0.04810127
#> 157    844.312 0.03164557
#> 158    847.688 0.04303797
#> 159    851.061 0.04556962
#> 160    854.433 0.04556962
#> 161    857.806 0.06329114
#> 162    861.175 0.04936709
#> 163    864.543 0.03417722
#> 164    867.910 0.05316456
#> 165    871.278 0.03417722
#> 166    874.642 0.03164557
#> 167    878.005 0.03924051
#> 168    881.367 0.03417722
#> 169    884.728 0.04683544
#> 170    888.087 0.04303797
#> 171    891.445 0.06075949
#> 172    894.802 0.04177215
#> 173    898.157 0.05949367
#> 174    901.511 0.03164557
#> 175    904.864 0.06075949
#> 176    908.216 0.04936709
#> 177    911.567 0.04810127
#> 178    914.914 0.04177215
#> 179    918.262 0.05569620
#> 180    921.609 0.04430380
#> 181    924.954 0.02025316
#> 182    928.296 0.05696203
#> 183    931.639 0.05063291
#> 184    934.981 0.03291139
#> 185    938.320 0.05569620
#> 186    941.659 0.05189873
#> 187    944.995 0.04050633
#> 188    948.332 0.04936709
#> 189    951.665 0.04810127
#> 190    954.999 0.05063291
#> 191    958.330 0.06202532
#> 192    961.660 0.04177215
#> 193    964.991 0.04303797
#> 194    968.318 0.04810127
#> 195    971.644 0.04683544
#> 196    974.971 0.05316456
#> 197    978.295 0.04430380
#> 198    981.617 0.03417722
#> 199    984.940 0.04303797
#> 200    988.258 0.05189873
#> 201    991.578 0.03924051
#> 202    994.896 0.06075949
#> 203    998.212 0.04177215
#> 204   1001.530 0.04810127
#> 205   1004.840 0.05443038
#> 206   1008.150 0.06329114
#> 207   1011.460 0.04177215
#> 208   1014.770 0.04936709
#> 209   1018.080 0.04936709
#> 210   1021.390 0.06202532
#> 211   1024.700 0.07594937
#> 212   1028.000 0.06582278
#> 213   1031.300 0.05569620
#> 214   1034.610 0.06329114
#> 215   1037.910 0.07341772
#> 216   1041.210 0.06582278
#> 217   1044.510 0.05443038
#> 218   1047.800 0.05063291
#> 219   1051.100 0.07215190
#> 220   1054.390 0.08101266
#> 221   1057.680 0.08860759
#> 222   1060.980 0.12911392
#> 223   1064.270 0.14810127
#> 224   1067.560 0.09493671
#> 225   1070.840 0.06962025
#> 226   1074.130 0.06329114
#> 227   1077.420 0.05822785
#> 228   1080.700 0.07468354
#> 229   1083.980 0.06708861
#> 230   1087.270 0.07215190
#> 231   1090.550 0.05189873
#> 232   1093.830 0.06708861
#> 233   1097.100 0.04810127
#> 234   1100.380 0.06835443
#> 235   1103.650 0.05822785
#> 236   1106.930 0.06075949
#> 237   1110.200 0.05569620
#> 238   1113.470 0.06962025
#> 239   1116.740 0.06708861
#> 240   1120.010 0.07088608
#> 241   1123.280 0.08734177
#> 242   1126.550 0.11772152
#> 243   1129.810 0.15443038
#> 244   1133.080 0.08607595
#> 245   1136.340 0.06835443
#> 246   1139.600 0.05696203
#> 247   1142.860 0.07088608
#> 248   1146.120 0.06329114
#> 249   1149.370 0.06329114
#> 250   1152.630 0.06329114
#> 251   1155.890 0.07088608
#> 252   1159.140 0.05949367
#> 253   1162.390 0.05063291
#> 254   1165.640 0.06835443
#> 255   1168.890 0.06708861
#> 256   1172.140 0.09493671
#> 257   1175.390 0.05696203
#> 258   1178.640 0.07215190
#> 259   1181.880 0.06202532
#> 260   1185.120 0.05063291
#> 261   1188.370 0.05822785
#> 262   1191.610 0.06329114
#> 263   1194.850 0.05316456
#> 264   1198.090 0.07848101
#> 265   1201.320 0.06962025
#> 266   1204.560 0.05063291
#> 267   1207.790 0.05189873
#> 268   1211.030 0.05316456
#> 269   1214.260 0.06962025
#> 270   1217.490 0.05189873
#> 271   1220.720 0.05443038
#> 272   1223.950 0.06582278
#> 273   1227.180 0.05316456
#> 274   1230.400 0.04810127
#> 275   1233.630 0.06962025
#> 276   1236.850 0.07215190
#> 277   1240.070 0.06582278
#> 278   1243.300 0.06075949
#> 279   1246.520 0.05189873
#> 280   1249.740 0.04810127
#> 281   1252.950 0.07088608
#> 282   1256.170 0.05822785
#> 283   1259.380 0.06708861
#> 284   1262.600 0.04936709
#> 285   1265.810 0.07341772
#> 286   1269.020 0.06329114
#> 287   1272.230 0.07721519
#> 288   1275.440 0.06202532
#> 289   1278.650 0.06329114
#> 290   1281.850 0.07721519
#> 291   1285.060 0.07088608
#> 292   1288.260 0.09493671
#> 293   1291.460 0.12911392
#> 294   1294.670 0.22025316
#> 295   1297.870 0.14810127
#> 296   1301.070 0.10000000
#> 297   1304.260 0.09493671
#> 298   1307.460 0.07721519
#> 299   1310.660 0.06455696
#> 300   1313.850 0.07848101
#> 301   1317.040 0.06329114
#> 302   1320.230 0.06708861
#> 303   1323.420 0.08734177
#> 304   1326.610 0.06329114
#> 305   1329.800 0.05189873
#> 306   1332.990 0.07088608
#> 307   1336.170 0.07088608
#> 308   1339.360 0.05569620
#> 309   1342.540 0.07088608
#> 310   1345.720 0.05822785
#> 311   1348.900 0.07848101
#> 312   1352.080 0.05063291
#> 313   1355.260 0.05316456
#> 314   1358.440 0.05696203
#> 315   1361.610 0.07341772
#> 316   1364.790 0.06075949
#> 317   1367.960 0.09367089
#> 318   1371.130 0.07215190
#> 319   1374.300 0.05443038
#> 320   1377.470 0.08227848
#> 321   1380.640 0.06962025
#> 322   1383.810 0.06455696
#> 323   1386.970 0.07341772
#> 324   1390.140 0.06202532
#> 325   1393.300 0.07594937
#> 326   1396.470 0.05696203
#> 327   1399.630 0.08354430
#> 328   1402.790 0.08607595
#> 329   1405.950 0.06455696
#> 330   1409.100 0.08481013
#> 331   1412.260 0.08481013
#> 332   1415.410 0.10632911
#> 333   1418.570 0.10886076
#> 334   1421.720 0.09746835
#> 335   1424.870 0.10126582
#> 336   1428.020 0.10632911
#> 337   1431.170 0.12531646
#> 338   1434.320 0.12278481
#> 339   1437.470 0.18101266
#> 340   1440.610 0.19240506
#> 341   1443.760 0.17721519
#> 342   1446.900 0.13037975
#> 343   1450.040 0.11518987
#> 344   1453.180 0.11392405
#> 345   1456.320 0.12784810
#> 346   1459.460 0.13924051
#> 347   1462.590 0.14810127
#> 348   1465.730 0.12784810
#> 349   1468.870 0.11139241
#> 350   1472.000 0.09493671
#> 351   1475.130 0.07468354
#> 352   1478.260 0.08481013
#> 353   1481.400 0.09113924
#> 354   1484.520 0.07721519
#> 355   1487.650 0.08987342
#> 356   1490.780 0.08607595
#> 357   1493.900 0.08734177
#> 358   1497.030 0.07468354
#> 359   1500.150 0.08607595
#> 360   1503.270 0.08481013
#> 361   1506.390 0.07468354
#> 362   1509.510 0.08481013
#> 363   1512.630 0.06075949
#> 364   1515.740 0.06202532
#> 365   1518.860 0.06962025
#> 366   1521.970 0.10000000
#> 367   1525.090 0.08734177
#> 368   1528.200 0.10000000
#> 369   1531.310 0.10000000
#> 370   1534.420 0.09746835
#> 371   1537.530 0.07594937
#> 372   1540.640 0.07468354
#> 373   1543.740 0.09620253
#> 374   1546.850 0.06202532
#> 375   1549.950 0.05822785
#> 376   1553.050 0.07974684
#> 377   1556.160 0.08860759
#> 378   1559.260 0.08354430
#> 379   1562.360 0.07594937
#> 380   1565.450 0.07468354
#> 381   1568.550 0.08101266
#> 382   1571.650 0.09367089
#> 383   1574.740 0.08227848
#> 384   1577.830 0.07848101
#> 385   1580.920 0.09113924
#> 386   1584.020 0.07215190
#> 387   1587.110 0.07341772
#> 388   1590.190 0.08481013
#> 389   1593.280 0.08987342
#> 390   1596.370 0.05696203
#> 391   1599.450 0.05949367
#> 392   1602.540 0.07848101
#> 393   1605.620 0.06582278
#> 394   1608.700 0.07341772
#> 395   1611.780 0.08101266
#> 396   1614.860 0.08987342
#> 397   1617.940 0.07974684
#> 398   1621.020 0.07848101
#> 399   1624.090 0.07594937
#> 400   1627.170 0.07974684
#> 401   1630.240 0.07974684
#> 402   1633.310 0.07974684
#> 403   1636.380 0.07974684
#> 404   1639.450 0.08860759
#> 405   1642.520 0.09113924
#> 406   1645.590 0.09746835
#> 407   1648.650 0.08607595
#> 408   1651.720 0.08227848
#> 409   1654.780 0.09113924
#> 410   1657.850 0.10379747
#> 411   1660.910 0.06835443
#> 412   1663.970 0.09367089
#> 413   1667.030 0.07848101
#> 414   1670.090 0.09493671
#> 415   1673.150 0.09873418
#> 416   1676.200 0.08227848
#> 417   1679.260 0.08860759
#> 418   1682.310 0.09113924
#> 419   1685.360 0.09620253
#> 420   1688.410 0.08860759
#> 421   1691.460 0.09367089
#> 422   1694.510 0.07088608
#> 423   1697.560 0.07721519
#> 424   1700.610 0.10253165
#> 425   1703.650 0.08607595
#> 426   1706.700 0.10000000
#> 427   1709.740 0.08860759
#> 428   1712.780 0.10000000
#> 429   1715.830 0.10126582
#> 430   1718.860 0.10632911
#> 431   1721.900 0.08860759
#> 432   1724.940 0.07848101
#> 433   1727.980 0.09113924
#> 434   1731.010 0.08607595
#> 435   1734.050 0.09240506
#> 436   1737.080 0.07468354
#> 437   1740.110 0.09240506
#> 438   1743.140 0.08481013
#> 439   1746.170 0.08227848
#> 440   1749.200 0.08860759
#> 441   1752.230 0.08607595
#> 442   1755.250 0.09746835
#> 443   1758.280 0.09367089
#> 444   1761.300 0.10126582
#> 445   1764.330 0.07594937
#> 446   1767.350 0.10000000
#> 447   1770.370 0.09367089
#> 448   1773.390 0.08101266
#> 449   1776.410 0.09620253
#> 450   1779.420 0.09367089
#> 451   1782.440 0.09367089
#> 452   1785.460 0.12025316
#> 453   1788.470 0.12025316
#> 454   1791.480 0.10253165
#> 455   1794.490 0.08987342
#> 456   1797.500 0.08607595
#> 457   1800.510 0.11392405
#> 458   1803.520 0.10000000
#> 459   1806.530 0.09367089
#> 460   1809.530 0.09113924
#> 461   1812.540 0.09620253
#> 462   1815.540 0.10253165
#> 463   1818.540 0.10632911
#> 464   1821.540 0.09367089
#> 465   1824.550 0.10253165
#> 466   1827.550 0.09493671
#> 467   1830.540 0.11139241
#> 468   1833.540 0.10126582
#> 469   1836.540 0.10506329
#> 470   1839.530 0.11265823
#> 471   1842.520 0.11392405
#> 472   1845.510 0.10379747
#> 473   1848.510 0.11139241
#> 474   1851.500 0.10886076
#> 475   1854.490 0.10632911
#> 476   1857.480 0.10886076
#> 477   1860.460 0.10379747
#> 478   1863.450 0.10253165
#> 479   1866.430 0.10506329
#> 480   1869.410 0.10506329
#> 481   1872.400 0.10000000
#> 482   1875.380 0.09620253
#> 483   1878.360 0.10253165
#> 484   1881.340 0.10759494
#> 485   1884.320 0.10506329
#> 486   1887.290 0.10379747
#> 487   1890.270 0.10886076
#> 488   1893.240 0.10379747
#> 489   1896.220 0.09620253
#> 490   1899.190 0.10000000
#> 491   1902.160 0.10632911
#> 492   1905.130 0.10379747
#> 493   1908.100 0.08734177
#> 494   1911.070 0.10253165
#> 495   1914.030 0.09240506
#> 496   1917.000 0.10126582
#> 497   1919.960 0.10632911
#> 498   1922.930 0.11265823
#> 499   1925.890 0.10253165
#> 500   1928.850 0.09367089
#> 501   1931.810 0.10379747
#> 502   1934.770 0.10253165
#> 503   1937.430 0.10253165
#> 504   1940.400 0.11012658
#> 505   1943.360 0.08481013
#> 506   1946.330 0.11012658
#> 507   1949.300 0.12405063
#> 508   1952.270 0.11772152
#> 509   1955.230 0.10506329
#> 510   1958.190 0.09493671
#> 511   1961.160 0.10506329
#> 512   1964.120 0.11898734
#> 513   1967.080 0.10886076
#> 514   1970.040 0.10379747
#> 515   1972.990 0.10759494
#> 516   1975.950 0.11392405
#> 517   1978.910 0.08987342
#> 518   1981.860 0.10126582
#> 519   1984.820 0.11139241
#> 520   1987.770 0.11392405
#> 521   1990.720 0.12531646
#> 522   1993.670 0.10506329
#> 523   1996.620 0.09746835
#> 524   1999.570 0.11139241
#> 525   2002.520 0.12151899
#> 526   2005.460 0.12151899
#> 527   2008.410 0.10126582
#> 528   2011.350 0.12405063
#> 529   2014.290 0.10000000
#> 530   2017.240 0.11898734
#> 531   2020.180 0.10000000
#> 532   2023.120 0.11645570
#> 533   2026.050 0.11139241
#> 534   2028.990 0.10000000
#> 535   2031.930 0.11645570
#> 536   2034.870 0.11265823
#> 537   2037.800 0.10886076
#> 538   2040.730 0.11392405
#> 539   2043.660 0.10886076
#> 540   2046.600 0.13417722
#> 541   2049.520 0.11392405
#> 542   2052.450 0.10126582
#> 543   2055.380 0.11265823
#> 544   2058.310 0.11012658
#> 545   2061.230 0.10379747
#> 546   2064.160 0.13924051
#> 547   2067.080 0.13291139
#> 548   2070.000 0.12278481
#> 549   2072.930 0.10379747
#> 550   2075.850 0.11898734
#> 551   2078.770 0.13291139
#> 552   2081.680 0.12025316
#> 553   2084.600 0.12658228
#> 554   2087.520 0.12405063
#> 555   2090.430 0.14050633
#> 556   2093.340 0.13164557
#> 557   2096.260 0.13037975
#> 558   2099.170 0.11772152
#> 559   2102.080 0.12658228
#> 560   2104.990 0.13924051
#> 561   2107.900 0.13797468
#> 562   2110.810 0.13797468
#> 563   2113.710 0.13164557
#> 564   2116.620 0.13164557
#> 565   2119.520 0.12405063
#> 566   2122.430 0.12658228
#> 567   2125.330 0.12151899
#> 568   2128.230 0.13291139
#> 569   2131.130 0.11898734
#> 570   2134.030 0.14430380
#> 571   2136.930 0.11898734
#> 572   2139.820 0.14177215
#> 573   2142.720 0.12911392
#> 574   2145.610 0.12784810
#> 575   2148.510 0.13291139
#> 576   2151.400 0.15189873
#> 577   2154.290 0.13544304
#> 578   2157.180 0.14556962
#> 579   2160.070 0.12405063
#> 580   2162.960 0.12278481
#> 581   2165.850 0.14050633
#> 582   2168.730 0.13291139
#> 583   2171.620 0.12784810
#> 584   2174.500 0.12911392
#> 585   2177.390 0.12278481
#> 586   2180.270 0.13797468
#> 587   2183.150 0.12278481
#> 588   2186.030 0.12025316
#> 589   2188.910 0.12911392
#> 590   2191.790 0.14556962
#> 591   2194.670 0.13924051
#> 592   2197.540 0.13924051
#> 593   2200.420 0.11772152
#> 594   2203.290 0.11139241
#> 595   2206.160 0.12151899
#> 596   2209.030 0.12784810
#> 597   2211.900 0.13797468
#> 598   2214.770 0.14177215
#> 599   2217.640 0.12658228
#> 600   2220.510 0.12531646
#> 601   2223.380 0.11645570
#> 602   2226.240 0.12025316
#> 603   2229.110 0.13037975
#> 604   2231.970 0.13544304
#> 605   2234.830 0.14050633
#> 606   2237.690 0.13291139
#> 607   2240.550 0.11645570
#> 608   2243.410 0.14683544
#> 609   2246.270 0.09746835
#> 610   2249.130 0.11012658
#> 611   2251.980 0.12278481
#> 612   2254.840 0.10632911
#> 613   2257.690 0.13797468
#> 614   2260.540 0.12911392
#> 615   2263.390 0.13544304
#> 616   2266.250 0.12025316
#> 617   2269.100 0.13291139
#> 618   2271.940 0.11645570
#> 619   2274.790 0.11265823
#> 620   2277.640 0.12151899
#> 621   2280.490 0.12405063
#> 622   2283.330 0.11139241
#> 623   2286.170 0.11898734
#> 624   2289.020 0.13924051
#> 625   2291.860 0.13037975
#> 626   2294.700 0.12278481
#> 627   2297.540 0.11772152
#> 628   2300.380 0.11772152
#> 629   2303.210 0.11645570
#> 630   2306.050 0.11139241
#> 631   2308.890 0.11392405
#> 632   2311.720 0.11392405
#> 633   2314.550 0.11898734
#> 634   2317.390 0.10379747
#> 635   2320.220 0.11012658
#> 636   2323.050 0.11139241
#> 637   2325.880 0.13037975
#> 638   2328.700 0.11392405
#> 639   2331.530 0.10632911
#> 640   2334.360 0.10506329
#> 641   2337.180 0.11645570
#> 642   2340.010 0.09746835
#> 643   2342.830 0.12025316
#> 644   2345.650 0.11898734
#> 645   2348.470 0.10379747
#> 646   2351.290 0.10379747
#> 647   2354.110 0.11772152
#> 648   2356.930 0.13164557
#> 649   2359.750 0.09873418
#> 650   2362.560 0.10253165
#> 651   2365.380 0.08734177
#> 652   2368.190 0.10632911
#> 653   2371.000 0.08860759
#> 654   2373.820 0.11518987
#> 655   2376.630 0.10506329
#> 656   2379.440 0.10759494
#> 657   2382.250 0.10759494
#> 658   2385.050 0.10886076
#> 659   2387.860 0.11012658
#> 660   2390.670 0.10886076
#> 661   2393.470 0.09493671
#> 662   2396.280 0.10632911
#> 663   2399.080 0.10253165
#> 664   2401.880 0.08860759
#> 665   2404.680 0.09240506
#> 666   2407.480 0.10000000
#> 667   2410.280 0.09493671
#> 668   2413.080 0.09240506
#> 669   2415.870 0.09746835
#> 670   2418.670 0.08987342
#> 671   2421.460 0.08860759
#> 672   2424.260 0.10379747
#> 673   2427.050 0.10253165
#> 674   2429.840 0.09746835
#> 675   2432.630 0.09746835
#> 676   2435.420 0.10886076
#> 677   2438.210 0.09746835
#> 678   2441.000 0.11265823
#> 679   2443.790 0.10379747
#> 680   2446.570 0.10253165
#> 681   2449.360 0.09620253
#> 682   2452.140 0.10379747
#> 683   2454.920 0.09746835
#> 684   2457.700 0.09620253
#> 685   2460.480 0.09113924
#> 686   2463.260 0.09620253
#> 687   2466.040 0.09113924
#> 688   2468.820 0.10000000
#> 689   2471.600 0.10253165
#> 690   2474.370 0.08734177
#> 691   2477.150 0.09746835
#> 692   2479.920 0.10886076
#> 693   2482.690 0.10886076
#> 694   2485.470 0.10000000
#> 695   2488.240 0.09240506
#> 696   2491.010 0.10000000
#> 697   2493.770 0.10886076
#> 698   2496.540 0.09240506
#> 699   2499.310 0.09240506
#> 700   2502.070 0.10126582
#> 701   2504.840 0.09493671
#> 702   2507.600 0.08987342
#> 703   2510.370 0.09620253
#> 704   2513.130 0.09240506
#> 705   2515.890 0.09240506
#> 706   2518.650 0.09620253
#> 707   2521.410 0.08734177
#> 708   2524.160 0.10506329
#> 709   2526.920 0.08860759
#> 710   2529.680 0.09367089
#> 711   2532.430 0.09240506
#> 712   2535.190 0.09746835
#> 713   2537.940 0.11518987
#> 714   2540.690 0.09367089
#> 715   2543.440 0.10126582
#> 716   2546.190 0.10632911
#> 717   2548.940 0.08860759
#> 718   2551.690 0.08987342
#> 719   2554.440 0.10000000
#> 720   2557.180 0.09113924
#> 721   2559.930 0.09367089
#> 722   2562.670 0.08101266
#> 723   2565.420 0.08607595
#> 724   2568.160 0.10253165
#> 725   2570.900 0.09493671
#> 726   2573.640 0.08354430
#> 727   2576.380 0.08860759
#> 728   2579.110 0.10126582
#> 729   2581.850 0.10000000
#> 730   2584.590 0.06329114
#> 731   2587.320 0.08481013
#> 732   2590.060 0.09873418
#> 733   2592.790 0.09493671
#> 734   2595.520 0.09873418
#> 735   2598.250 0.08734177
#> 736   2600.980 0.08734177
#> 737   2603.720 0.08481013
#> 738   2606.440 0.06962025
#> 739   2609.170 0.09873418
#> 740   2611.900 0.08860759
#> 741   2614.620 0.09493671
#> 742   2617.350 0.09113924
#> 743   2620.070 0.08227848
#> 744   2622.790 0.09746835
#> 745   2625.510 0.07468354
#> 746   2628.230 0.09367089
#> 747   2630.950 0.07721519
#> 748   2633.670 0.07721519
#> 749   2636.390 0.09113924
#> 750   2639.110 0.09367089
#> 751   2641.820 0.09367089
#> 752   2644.540 0.08734177
#> 753   2647.250 0.11139241
#> 754   2649.960 0.08860759
#> 755   2652.680 0.10000000
#> 756   2655.390 0.10126582
#> 757   2658.100 0.06962025
#> 758   2660.810 0.08860759
#> 759   2663.510 0.09873418
#> 760   2666.220 0.09746835
#> 761   2668.930 0.07974684
#> 762   2671.630 0.10379747
#> 763   2674.340 0.08354430
#> 764   2677.040 0.07974684
#> 765   2679.740 0.08860759
#> 766   2682.440 0.06962025
#> 767   2685.140 0.08227848
#> 768   2687.840 0.09367089
#> 769   2690.540 0.09620253
#> 770   2693.240 0.09746835
#> 771   2695.940 0.09240506
#> 772   2698.630 0.11772152
#> 773   2701.330 0.09873418
#> 774   2704.020 0.08987342
#> 775   2706.710 0.09240506
#> 776   2709.410 0.12911392
#> 777   2712.100 0.11392405
#> 778   2714.790 0.12658228
#> 779   2717.470 0.12278481
#> 780   2720.160 0.13164557
#> 781   2722.850 0.12025316
#> 782   2725.530 0.13417722
#> 783   2728.220 0.11265823
#> 784   2730.900 0.10759494
#> 785   2733.590 0.10506329
#> 786   2736.270 0.09620253
#> 787   2738.950 0.08860759
#> 788   2741.630 0.10379747
#> 789   2744.310 0.09873418
#> 790   2746.990 0.07088608
#> 791   2749.670 0.09367089
#> 792   2752.340 0.10253165
#> 793   2755.020 0.08987342
#> 794   2757.690 0.06962025
#> 795   2760.370 0.09367089
#> 796   2763.040 0.09493671
#> 797   2765.710 0.09113924
#> 798   2768.380 0.09113924
#> 799   2771.050 0.09367089
#> 800   2773.720 0.09620253
#> 801   2776.390 0.08987342
#> 802   2779.060 0.08734177
#> 803   2781.720 0.08101266
#> 804   2784.390 0.08481013
#> 805   2787.050 0.09620253
#> 806   2789.710 0.08734177
#> 807   2792.380 0.09746835
#> 808   2795.040 0.10126582
#> 809   2797.700 0.09620253
#> 810   2800.360 0.09240506
#> 811   2803.020 0.09493671
#> 812   2805.680 0.11518987
#> 813   2808.330 0.09493671
#> 814   2810.990 0.09873418
#> 815   2813.640 0.11139241
#> 816   2816.300 0.10886076
#> 817   2818.950 0.11772152
#> 818   2821.600 0.10126582
#> 819   2824.250 0.11392405
#> 820   2826.900 0.13291139
#> 821   2829.550 0.13417722
#> 822   2832.200 0.15443038
#> 823   2834.850 0.18101266
#> 824   2837.490 0.22025316
#> 825   2840.140 0.35063291
#> 826   2842.780 0.44556962
#> 827   2845.430 0.53924051
#> 828   2848.070 0.58607595
#> 829   2850.710 0.55569620
#> 830   2853.350 0.51265823
#> 831   2855.990 0.45822785
#> 832   2858.630 0.41772152
#> 833   2861.270 0.38860759
#> 834   2863.910 0.36962025
#> 835   2866.540 0.35316456
#> 836   2869.180 0.35696203
#> 837   2871.810 0.35316456
#> 838   2874.450 0.39113924
#> 839   2877.080 0.48227848
#> 840   2879.710 0.69620253
#> 841   2882.340 1.00000000
#> 842   2884.970 0.80759494
#> 843   2887.600 0.59240506
#> 844   2890.230 0.46455696
#> 845   2892.860 0.41265823
#> 846   2895.480 0.36962025
#> 847   2898.110 0.37468354
#> 848   2900.730 0.37341772
#> 849   2903.350 0.35696203
#> 850   2905.970 0.32025316
#> 851   2908.600 0.28860759
#> 852   2911.220 0.29240506
#> 853   2913.840 0.26835443
#> 854   2916.460 0.24810127
#> 855   2919.070 0.23797468
#> 856   2921.690 0.24430380
#> 857   2924.310 0.21898734
#> 858   2926.920 0.22911392
#> 859   2929.540 0.23544304
#> 860   2932.150 0.25316456
#> 861   2934.760 0.22784810
#> 862   2937.370 0.19620253
#> 863   2939.980 0.20253165
#> 864   2942.590 0.17215190
#> 865   2945.200 0.15443038
#> 866   2947.810 0.10253165
#> 867   2950.420 0.11772152
#> 868   2953.020 0.10632911
#> 869   2955.630 0.11012658
#> 870   2958.230 0.09493671
#> 871   2960.830 0.10506329
#> 872   2963.440 0.11139241
#> 873   2966.040 0.10000000
#> 874   2968.640 0.10253165
#> 875   2971.240 0.08607595
#> 876   2973.830 0.08227848
#> 877   2976.430 0.08987342
#> 878   2979.030 0.09493671
#> 879   2981.630 0.07721519
#> 880   2984.220 0.10379747
#> 881   2986.810 0.08734177
#> 882   2989.410 0.09746835
#> 883   2992.000 0.08734177
#> 884   2994.590 0.08607595
#> 885   2997.180 0.08860759
#> 886   2999.770 0.07974684
#> 887   3002.360 0.07974684
#> 888   3004.950 0.10000000
#> 889   3007.530 0.08607595
#> 890   3010.120 0.06708861
#> 891   3012.700 0.07468354
#> 892   3015.290 0.09240506
#> 893   3017.870 0.08481013
#> 894   3020.450 0.07341772
#> 895   3023.040 0.06455696
#> 896   3025.620 0.08734177
#> 897   3028.200 0.08354430
#> 898   3030.770 0.05949367
#> 899   3033.350 0.07974684
#> 900   3035.930 0.09113924
#> 901   3038.500 0.05822785
#> 902   3041.080 0.06835443
#> 903   3043.650 0.06962025
#> 904   3046.230 0.09113924
#> 905   3048.800 0.07721519
#> 906   3051.370 0.06962025
#> 907   3053.940 0.07594937
#> 908   3056.510 0.08354430
#> 909   3059.080 0.07468354
#> 910   3061.650 0.07594937
#> 911   3064.210 0.08354430
#> 912   3066.780 0.07594937
#> 913   3069.350 0.07088608
#> 914   3071.910 0.07721519
#> 915   3074.470 0.07594937
#> 916   3077.040 0.08481013
#> 917   3079.600 0.07721519
#> 918   3082.160 0.07848101
#> 919   3084.720 0.07721519
#> 920   3087.280 0.06962025
#> 921   3089.830 0.07088608
#> 922   3092.390 0.08101266
#> 923   3094.950 0.06329114
#> 924   3097.500 0.08101266
#> 925   3100.060 0.07215190
#> 926   3102.610 0.06835443
#> 927   3105.160 0.06202532
#> 928   3107.720 0.07088608
#> 929   3110.270 0.07215190
#> 930   3112.820 0.06329114
#> 931   3115.370 0.06075949
#> 932   3117.910 0.07341772
#> 933   3120.460 0.07215190
#> 934   3123.010 0.05949367
#> 935   3125.560 0.06075949
#> 936   3128.100 0.06455696
#> 937   3130.640 0.06455696
#> 938   3133.190 0.06455696
#> 939   3135.730 0.06962025
#> 940   3138.270 0.06835443
#> 941   3140.810 0.06455696
#> 942   3143.350 0.06202532
#> 943   3145.890 0.05316456
#> 944   3148.430 0.05443038
#> 945   3150.960 0.06455696
#> 946   3153.500 0.06455696
#> 947   3156.030 0.05949367
#> 948   3158.570 0.06075949
#> 949   3161.100 0.06962025
#> 950   3163.640 0.06202532
#> 951   3166.170 0.05316456
#> 952   3168.700 0.05949367
#> 953   3171.230 0.06202532
#> 954   3173.760 0.06329114
#> 955   3175.300 0.06455696
#> 956   3177.840 0.06202532
#> 957   3180.380 0.06455696
#> 958   3182.910 0.05696203
#> 959   3185.450 0.06455696
#> 960   3187.990 0.05696203
#> 961   3190.520 0.05696203
#> 962   3193.060 0.06202532
#> 963   3195.590 0.06202532
#> 964   3198.120 0.05189873