Match spectra with reference library

match_spec() will compare a spectrum to a spectral library formatted with the Open Specy standard and report the best match using the Pearson correlation coefficient. find_spec() makes it easy to retrieve single spectra and metadata from the Open Specy reference library.

Usage

correlate_spectra(data, library, ...)

Arguments

data

a data frame containing the variables in formula.

library

reference library you want to compare against.

x

a numeric vector containing the spectral wavenumbers; alternatively a data frame containing spectral data as "wavenumber" and "intensity" can be supplied.

y

a numeric vector containing the spectral intensities.

formula

an object of class 'formula' of the form intensity ~ wavenumber.

which

a character string specifying which library to match, "raman" or "ftir".

type

a character string specifying whether the "full" spectrum should be matched or spectrum "peaks" only. "metadata" is needed to browser spectra with find_spec().

range

this should be all possible wavenumber values from your spectral library.

top_n

number of top matches that you want to be returned.

subset

logical expression indicating elements or rows to search for; see subset() for details.

cols

columns to retrieve from the Open Specy reference library; columns containing no or missing values are automatically removed.

...

further arguments passed to the submethods.

Value

match_spec() returns a data frame with the top_n material matches, their Pearson's r value, and the organization they were provided by. find_spec() returns a data frame with the spectral raw data or metadata of a specific reference spectrum.

Details

This routine will match the spectrum you want to identify to the wavenumbers present in the spectral library. Once the spectra are aligned, it computes the Pearson correlation coefficient between the spectrum you want to identify and all spectra in the library (see cor). The function returns a table with the Pearson correlation coefficient values and all metadata for the top spectral matches. If using the Open Specy library, all intensity values are in absorbance, so your spectra should also be in absorbance units. If you need to convert your spectrum, use adj_intens().

See also

Author

Win Cowger, Zacharias Steinmetz

Examples

if (FALSE) {
data("raman_hdpe")

get_lib("raman")
spec_lib <- load_lib("raman")

match_spec(raman_proc, library = spec_lib, which = "raman")

find_spec(sample_name == 5381, library = spec_lib, which = "raman")
}