Define features

Functions for analyzing features, like particles, fragments, or fibers, in spectral map oriented OpenSpecy object.

Usage

collapse_spec(x, ...)

# S3 method for default
collapse_spec(x, ...)

# S3 method for OpenSpecy
collapse_spec(x, ...)

def_features(x, ...)

# S3 method for default
def_features(x, ...)

# S3 method for OpenSpecy
def_features(x, features, ...)

Arguments

x

an OpenSpecy object

features

a logical vector or character vector describing which of the spectra are of features (TRUE) and which are not (FALSE). If a character vector is provided, it should represent the different feature types present in the spectra.

...

additional arguments passed to subfunctions.

Value

An OpenSpecy object appended with metadata about the features or collapsed for the features.

Details

def_features() accepts an OpenSpecy object and a logical or character vector describing which pixels correspond to particles. collapse_spec() takes an OpenSpecy object with particle-specific metadata (from def_features()) and collapses the spectra to median intensities for each unique particle. It also updates the metadata with centroid coordinates, while preserving the feature information on area and Feret max.

Author

Win Cowger, Zacharias Steinmetz

Examples

data.table::setDTthreads(2)
tiny_map <- read_extdata("CA_tiny_map.zip") |> read_any()
identified_map <- def_features(tiny_map, tiny_map$metadata$x == 0)
collapse_spec(identified_map)
#>      wavenumber          0        -88          4          7         10
#>   1:     717.42 -0.8599975 -0.1846190 -0.7817467 -0.6341125 -0.3957456
#>   2:     725.13 -0.8629220 -0.2030258 -0.7875991 -0.6482702 -0.4134940
#>   3:     732.85 -0.8684751 -0.2236059 -0.7965024 -0.6606091 -0.4314153
#>   4:     740.56 -0.8713333 -0.2380186 -0.8020197 -0.6671489 -0.4433241
#>   5:     748.28 -0.8721815 -0.2436821 -0.8036770 -0.6697885 -0.4486529
#>  ---                                                                  
#> 423:    3972.81 -0.8804550 -0.3092059 -0.8022860 -0.6340107 -0.4140871
#> 424:    3980.52 -0.8789783 -0.3107601 -0.8013673 -0.6315157 -0.4108835
#> 425:    3988.24 -0.8772697 -0.3180229 -0.8017516 -0.6310282 -0.4105957
#> 426:    3995.95 -0.8773139 -0.3195935 -0.8025478 -0.6334830 -0.4125072
#> 427:    4003.67 -0.8778735 -0.3126730 -0.8019281 -0.6349432 -0.4132366
#>               13          3          6          9           12          2
#>   1: -0.06994937 0.11460015 0.18089953 0.17514743  0.087934248 -0.1404191
#>   2: -0.09679992 0.08463844 0.15293343 0.14219251  0.071245745 -0.1482290
#>   3: -0.12339281 0.05434874 0.12460253 0.11172757  0.051149037 -0.1594115
#>   4: -0.14147647 0.03276557 0.10344046 0.09271412  0.036638767 -0.1682616
#>   5: -0.15012543 0.02154485 0.08880223 0.08287866  0.032449901 -0.1715885
#>  ---                                                                     
#> 423: -0.10772412 0.03843024 0.13000481 0.14231284  0.003167340 -0.2771131
#> 424: -0.10918775 0.03988123 0.13281861 0.14542726  0.006487147 -0.2751649
#> 425: -0.10935027 0.03883109 0.13485433 0.14393683  0.006562222 -0.2740324
#> 426: -0.11035092 0.03590450 0.13375901 0.14145134  0.002249188 -0.2760093
#> 427: -0.11206047 0.03426958 0.13073599 0.14077842 -0.003012084 -0.2802055
#>               5          8         11
#>   1: -0.3454294 -0.4089794 -0.5889121
#>   2: -0.3505079 -0.4128581 -0.5879138
#>   3: -0.3595735 -0.4174158 -0.5898755
#>   4: -0.3662870 -0.4215689 -0.5909628
#>   5: -0.3689005 -0.4243726 -0.5905932
#>  ---                                 
#> 423: -0.4831747 -0.5125267 -0.6319057
#> 424: -0.4809514 -0.5105906 -0.6320986
#> 425: -0.4795221 -0.5097994 -0.6328617
#> 426: -0.4812161 -0.5112385 -0.6342685
#> 427: -0.4826814 -0.5125880 -0.6346731
#> 
#> $metadata
#>     x  y feature_id area perimeter feret_min feret_max       file_name  license
#>  1: 0  0          0    1         4         1         1 CA small UF.dat CC BY-NC
#>  2: 1  0        -88   NA        NA        NA        NA CA small UF.dat CC BY-NC
#>  3: 0  1          4    1         4         1         1 CA small UF.dat CC BY-NC
#>  4: 0  2          7    1         4         1         1 CA small UF.dat CC BY-NC
#>  5: 0  3         10    1         4         1         1 CA small UF.dat CC BY-NC
#>  6: 0  4         13    1         4         1         1 CA small UF.dat CC BY-NC
#>  7: 0  5          3    1         4         1         1 CA small UF.dat CC BY-NC
#>  8: 0  6          6    1         4         1         1 CA small UF.dat CC BY-NC
#>  9: 0  7          9    1         4         1         1 CA small UF.dat CC BY-NC
#> 10: 0  8         12    1         4         1         1 CA small UF.dat CC BY-NC
#> 11: 0  9          2    1         4         1         1 CA small UF.dat CC BY-NC
#> 12: 0 10          5    1         4         1         1 CA small UF.dat CC BY-NC
#> 13: 0 11          8    1         4         1         1 CA small UF.dat CC BY-NC
#> 14: 0 12         11    1         4         1         1 CA small UF.dat CC BY-NC
#>                                                      description samples lines
#>  1: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  2: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  3: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  4: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  5: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  6: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  7: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  8: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  9: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 10: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 11: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 12: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 13: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 14: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>     bands header offset data type interleave                  z plot titles
#>  1:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  2:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  3:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  4:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  5:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  6:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  7:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  8:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  9:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 10:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 11:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 12:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 13:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 14:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>                   pixel size col_id
#>  1: 0.000025000, 0.000025000    0_0
#>  2: 0.000025000, 0.000025000    0_1
#>  3: 0.000025000, 0.000025000    1_0
#>  4: 0.000025000, 0.000025000    2_0
#>  5: 0.000025000, 0.000025000    3_0
#>  6: 0.000025000, 0.000025000    4_0
#>  7: 0.000025000, 0.000025000    5_0
#>  8: 0.000025000, 0.000025000    6_0
#>  9: 0.000025000, 0.000025000    7_0
#> 10: 0.000025000, 0.000025000    8_0
#> 11: 0.000025000, 0.000025000    9_0
#> 12: 0.000025000, 0.000025000   10_0
#> 13: 0.000025000, 0.000025000   11_0
#> 14: 0.000025000, 0.000025000   12_0
#>                                                            session_id
#>  1: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  2: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  3: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  4: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  5: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  6: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  7: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  8: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>  9: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#> 10: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#> 11: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#> 12: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#> 13: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#> 14: 5f97c942a8f63d2c8062e9c15cdcddd3/8f227d2d59941c811cf7eb0672e5e2c2
#>                              file_id centroid_x centroid_y
#>  1: 215a0fd7e0dc3b1001799214618ce6eb          0          0
#>  2: 215a0fd7e0dc3b1001799214618ce6eb          8          6
#>  3: 215a0fd7e0dc3b1001799214618ce6eb          0          1
#>  4: 215a0fd7e0dc3b1001799214618ce6eb          0          2
#>  5: 215a0fd7e0dc3b1001799214618ce6eb          0          3
#>  6: 215a0fd7e0dc3b1001799214618ce6eb          0          4
#>  7: 215a0fd7e0dc3b1001799214618ce6eb          0          5
#>  8: 215a0fd7e0dc3b1001799214618ce6eb          0          6
#>  9: 215a0fd7e0dc3b1001799214618ce6eb          0          7
#> 10: 215a0fd7e0dc3b1001799214618ce6eb          0          8
#> 11: 215a0fd7e0dc3b1001799214618ce6eb          0          9
#> 12: 215a0fd7e0dc3b1001799214618ce6eb          0         10
#> 13: 215a0fd7e0dc3b1001799214618ce6eb          0         11
#> 14: 215a0fd7e0dc3b1001799214618ce6eb          0         12