Read spectral data

Functions for reading spectral data types including .asp, .jdx, .spc, .spa, .0, and .csv. Currently supported reading formats are .0, .asp, .spc, .csv, .spa, .zip, .yaml, .json, and .rds. If a zip file, the function supports any of the files listed above or additionally map files in .dat and .hdr format. Proprietary file formats like .0, .asp, and .spa are poorly supported but will likely still work in most cases.

Usage

read_any(file, share = NULL)

read_text(
  file,
  colnames = NULL,
  method = "fread",
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

read_asp(
  file,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

read_spa(
  file,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

read_jdx(
  file,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

read_spc(
  file,
  share = NULL,
  metadata = list(file_name = basename(file), user_name = NULL, contact_info = NULL,
    organization = NULL, citation = NULL, spectrum_type = NULL, spectrum_identity = NULL,
    material_form = NULL, material_phase = NULL, material_producer = NULL,
    material_purity = NULL, material_quality = NULL, material_color = NULL,
    material_other = NULL, cas_number = NULL, instrument_used = NULL,
    instrument_accessories = NULL, instrument_mode = NULL, spectral_resolution = NULL,
    laser_light_used = NULL, number_of_accumulations = NULL, 
    
    total_acquisition_time_s = NULL, data_processing_procedure = NULL,
    level_of_confidence_in_identification = NULL, other_info = NULL, license =
    "CC BY-NC"),
  ...
)

read_extdata(file = NULL)

Arguments

file

file to be read from or written to.

share

defaults to NULL; needed to share spectra with the Open Specy community; see share_spec() for details.

colnames

character vector of length = 2 indicating the colum names for the wavenumber and intensity; if NULL columns are guessed.

method

submethod to be used for reading text files; defaults to fread() but read.csv() works as well.

metadata

a named list of the metadata; see details below.

...

further arguments passed to the submethods.

Value

All read_*() functions return data frames containing two columns named "wavenumber" and "intensity".

Details

read_spc() and read_jdx() are just a wrapper around the functions provided by the hyperSpec package. read_opus() is a wrapper around opus_reader https://github.com/pierreroudier/opusreader. Other functions have been adapted various online sources. All reading functions convert datasets to a threepart list, one with a vector of the wavenumbers of the spectra, the second with a data.table of all spectral intensities ordered as columns, the third item is another data.table with any metadata the user provides or is harvested from the files themselves. Currently metadata harvesting from jdx and opus files are supported as well as the two Open Specy write formats yaml and json. There are many unique iterations of spectral file formats so there may be bugs in the file conversion. Please contact us if you identify any.

The metadata argument may contain a named list with the following details (* = minimum recommended):

file_name*:	The file name, defaults to basename() if not specified

user_name*:	User name, e.g. "Win Cowger"

contact_info:	Contact information, e.g. "1-513-673-8956, wincowger@gmail.com"

organization:	Affiliation, e.g. "University of California, Riverside"

citation:	Data citation, e.g. "Primpke, S., Wirth, M., Lorenz, C., & Gerdts, G. (2018). Reference database design for the automated analysis of microplastic samples based on Fourier transform infrared (FTIR) spectroscopy. Analytical and Bioanalytical Chemistry. doi:10.1007/s00216-018-1156-x "

spectrum_type*:	Raman or FTIR

spectrum_identity*:	Material/polymer analyzed, e.g. "Polystyrene"

material_form:	Form of the material analyzed, e.g. textile fiber, rubber band, sphere, granule

material_phase:	Phase of the material analyzed (liquid, gas, solid)

material_producer:	Producer of the material analyzed, e.g. Dow

material_purity:	Purity of the material analyzed, e.g. 99.98%

material_quality:	Quality of the material analyzed, e.g. consumer product, manufacturer material, analytical standard, environmental sample

material_color:	Color of the material analyzed, e.g. blue, #0000ff, (0, 0, 255)

material_other:	Other material description, e.g. 5 µm diameter fibers, 1 mm spherical particles

cas_number:	CAS number, e.g. 9003-53-6

instrument_used:	Instrument used, e.g. Horiba LabRam

instrument_accessories:	Instrument accessories, e.g. Focal Plane Array, CCD

instrument_mode:	Instrument modes/settings, e.g. transmission, reflectance

spectral_resolution:	Spectral resolution, e.g. 4/cm

laser_light_used:	Wavelength of the laser/light used, e.g. 785 nm

number_of_accumulations:	Number of accumulations, e.g 5

total_acquisition_time_s:	Total acquisition time (s), e.g. 10 s

data_processing_procedure:	Data processing procedure, e.g. spikefilter, baseline correction, none

level_of_confidence_in_identification:	Level of confidence in identification, e.g. 99%

other_info:	Other information

license:	The license of the shared spectrum; defaults to "CC BY-NC" (see https://creativecommons.org/licenses/by-nc/4.0/ for details). Any other creative commons license is allowed, for example, CC0 or CC BY

session_id:	A unique user and session identifier; populated automatically with paste(digest(Sys.info()), digest(sessionInfo()), sep = "/")

file_id:	A unique file identifier; populated automatically with digest(object[c("wavenumber", "spectra")])

See also

Author

Zacharias Steinmetz, Win Cowger

Examples

read_any(read_extdata("raman_hdpe.csv"))
#> Error in read_any(read_extdata("raman_hdpe.csv")): could not find function "read_any"
read_any(read_extdata("raman_atacamit.spc"))
#> Error in read_any(read_extdata("raman_atacamit.spc")): could not find function "read_any"
read_any(read_extdata("CA_tiny_map.zip"))
#> Error in read_any(read_extdata("CA_tiny_map.zip")): could not find function "read_any"
read_any(read_extdata("ftir_nitrocellulose.jdx"))
#> Error in read_any(read_extdata("ftir_nitrocellulose.jdx")): could not find function "read_any"
read_any(read_extdata("ftir_ldpe_soil.asp"))
#> Error in read_any(read_extdata("ftir_ldpe_soil.asp")): could not find function "read_any"
read_any(read_extdata("testdata_zipped.zip"))
#> Error in read_any(read_extdata("testdata_zipped.zip")): could not find function "read_any"
read_any(read_extdata("ftir_ps.0"))
#> Error in read_any(read_extdata("ftir_ps.0")): could not find function "read_any"