Read spectral data from multiple files

Wrapper functions for reading files in batch.

Usage

read_any(file, ...)

read_zip(file, ...)

Arguments

file

file to be read from or written to.

...

further arguments passed to the submethods.

Value

All read_*() functions return OpenSpecy objects

Details

read_any() provides a single function to quickly read in any of the supported formats, it assumes that the file extension will tell it how to process the spectra. read_zip() provides functionality for reading in spectral map files with ENVI file format or as individual files in a zip folder. If individual files, spectra are concatenated.

See also

Author

Zacharias Steinmetz, Win Cowger

Examples

read_extdata("raman_hdpe.csv") |> read_any()
#>      wavenumber intensity
#>   1:    301.040        26
#>   2:    304.632        50
#>   3:    308.221        48
#>   4:    311.810        45
#>   5:    315.398        46
#>  ---                     
#> 960:   3187.990        71
#> 961:   3190.520        71
#> 962:   3193.060        75
#> 963:   3195.590        75
#> 964:   3198.120        67
#> 
#> $metadata
#>    x y      file_name  license
#> 1: 1 1 raman_hdpe.csv CC BY-NC
#>                                                           session_id
#> 1: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>                             file_id    col_id
#> 1: cb06ce2846b119d932fb6696479a445b intensity
read_extdata("ftir_ldpe_soil.asp") |> read_any()
#>       wavenumber   intensity
#>    1:   650.4205 0.027056019
#>    2:   652.2841 0.029133933
#>    3:   654.1478 0.032566057
#>    4:   656.0115 0.032249778
#>    5:   657.8751 0.028957002
#>   ---                       
#> 1794:  3991.9788 0.005803670
#> 1795:  3993.8425 0.005495489
#> 1796:  3995.7062 0.005133253
#> 1797:  3997.5698 0.005595756
#> 1798:  3999.4335 0.005694630
#> 
#> $metadata
#>    x y          file_name  license
#> 1: 1 1 ftir_ldpe_soil.asp CC BY-NC
#>                                                           session_id
#> 1: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>                             file_id    col_id
#> 1: a33700bad0376ddab156980f473bac92 intensity
read_extdata("ftir_ps.0") |> read_any()
#>       wavenumber         V1
#>    1:   399.2239 0.01660202
#>    2:   401.1525 0.01790014
#>    3:   403.0811 0.01831616
#>    4:   405.0097 0.01894823
#>    5:   406.9384 0.02070733
#>   ---                      
#> 2122:  4489.8223 0.02349887
#> 2123:  4491.7509 0.02352615
#> 2124:  4493.6796 0.02357077
#> 2125:  4495.6082 0.02362563
#> 2126:  4497.5368 0.02361715
#> 
#> $metadata
#>    x y file_name  license                         unique_id     sample_id
#> 1: 1 1 ftir_ps.0 CC BY-NC sample1_PSnew_2020-10-28 14:33:02 sample1_PSnew
#>           date_time_sm        date_time_rf   sample_name instr_name_range
#> 1: 2020-10-28 14:33:02 2020-10-28 14:23:20 sample1_PSnew    VERTEX 70-MIR
#>    resolution_wn result_spc beamspl laser_wn
#> 1:             4         AB     KBr 15799.24
#>                               spc_in_file zero_filling temp_scanner_sm
#> 1: ig_sample;ig_ref;sc_sample;sc_ref;spec            2            27.9
#>    temp_scanner_rf hum_rel_sm hum_rel_rf hum_abs_sm hum_abs_rf
#> 1:            27.9          2         NA         NA         NA
#>                                                           session_id
#> 1: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>                             file_id col_id
#> 1: 0e9318d04fdb3beb4cd61145b080ccb6     V1
read_extdata("testdata_zipped.zip") |> read_zip()
#>      wavenumber intensity intensity.1 intensity.2
#>   1:    301.040        26          26          26
#>   2:    304.632        50          50          50
#>   3:    308.221        48          48          48
#>   4:    311.810        45          45          45
#>   5:    315.398        46          46          46
#>  ---                                             
#> 960:   3187.990        71          71          71
#> 961:   3190.520        71          71          71
#> 962:   3193.060        75          75          75
#> 963:   3195.590        75          75          75
#> 964:   3198.120        67          67          67
#> 
#> $metadata
#>    x y                file_name  license
#> 1: 1 1 testdata2 - Copy (2).csv CC BY-NC
#> 2: 2 1     testdata2 - Copy.csv CC BY-NC
#> 3: 1 2            testdata2.csv CC BY-NC
#>                                                           session_id
#> 1: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 2: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 3: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>                             file_id    col_id
#> 1: cb06ce2846b119d932fb6696479a445b intensity
#> 2: cb06ce2846b119d932fb6696479a445b intensity
#> 3: cb06ce2846b119d932fb6696479a445b intensity
read_extdata("CA_tiny_map.zip") |> read_zip()
#>      wavenumber        0_0       0_1        0_2       0_3        0_4       0_5
#>   1:     717.42 -0.8599975 -1.176549 -0.8269680 -1.150691 -0.7874242 -1.131251
#>   2:     725.13 -0.8629220 -1.184673 -0.8241877 -1.145349 -0.7848697 -1.132279
#>   3:     732.85 -0.8684751 -1.201088 -0.8249527 -1.145835 -0.7878844 -1.138330
#>   4:     740.56 -0.8713333 -1.215284 -0.8285120 -1.151216 -0.7958543 -1.145293
#>   5:     748.28 -0.8721815 -1.216835 -0.8334568 -1.157416 -0.8058372 -1.147371
#>  ---                                                                          
#> 423:    3972.81 -0.8804550 -1.293169 -0.8440830 -1.230502 -0.8018205 -1.190946
#> 424:    3980.52 -0.8789783 -1.295723 -0.8462818 -1.233429 -0.8030480 -1.190417
#> 425:    3988.24 -0.8772697 -1.305601 -0.8471817 -1.241330 -0.8022795 -1.192631
#> 426:    3995.95 -0.8773139 -1.310077 -0.8457789 -1.247904 -0.8011338 -1.193246
#> 427:    4003.67 -0.8778735 -1.303940 -0.8450245 -1.246171 -0.8017741 -1.188956
#>             0_6       0_7        0_8       0_9       0_10      0_11       0_12
#>   1: -0.8181180 -1.088998 -0.8292431 -1.072920 -0.8490217 -1.044781 -0.8055528
#>   2: -0.8157052 -1.094163 -0.8254585 -1.074778 -0.8455939 -1.042407 -0.8023769
#>   3: -0.8137981 -1.103946 -0.8287728 -1.076342 -0.8458889 -1.043109 -0.8070771
#>   4: -0.8150936 -1.110569 -0.8353058 -1.079404 -0.8477870 -1.045840 -0.8143305
#>   5: -0.8188222 -1.112912 -0.8405676 -1.084448 -0.8502600 -1.050660 -0.8213006
#>  ---                                                                          
#> 423: -0.7831531 -1.134469 -0.7993639 -1.119114 -0.8128311 -1.076820 -0.8029224
#> 424: -0.7858623 -1.138433 -0.7996801 -1.123678 -0.8118886 -1.079553 -0.8038837
#> 425: -0.7878010 -1.146599 -0.7993538 -1.128633 -0.8121727 -1.081921 -0.8041241
#> 426: -0.7882353 -1.147445 -0.7993204 -1.130919 -0.8131239 -1.084749 -0.8037490
#> 427: -0.7870936 -1.141112 -0.8001940 -1.132363 -0.8130202 -1.086768 -0.8026713
#>            0_13       0_14       0_15        1_0       1_1        1_2
#>   1: -0.9851461 -0.7893320 -0.8729300 -0.7817467 -1.037279 -0.6797068
#>   2: -0.9868538 -0.7853272 -0.8692073 -0.7875991 -1.048603 -0.6794843
#>   3: -0.9862352 -0.7823833 -0.8676148 -0.7965024 -1.067380 -0.6840024
#>   4: -0.9870666 -0.7851717 -0.8697423 -0.8020197 -1.083057 -0.6892709
#>   5: -0.9920748 -0.7919775 -0.8769495 -0.8036770 -1.085820 -0.6939027
#>  ---                                                                 
#> 423: -1.0179443 -0.7737747 -0.8834741 -0.8022860 -1.147436 -0.6932624
#> 424: -1.0177335 -0.7740544 -0.8812205 -0.8013673 -1.147939 -0.6938105
#> 425: -1.0158888 -0.7744763 -0.8804378 -0.8017516 -1.155403 -0.6935061
#> 426: -1.0132719 -0.7738099 -0.8825330 -0.8025478 -1.158695 -0.6918315
#> 427: -1.0116997 -0.7726955 -0.8838044 -0.8019281 -1.152465 -0.6914637
#>             1_3        1_4        1_5        1_6        1_7        1_8
#>   1: -0.9080905 -0.6090969 -0.8677020 -0.6418716 -0.8795404 -0.6977192
#>   2: -0.9057059 -0.6073114 -0.8766068 -0.6438789 -0.8903178 -0.6934057
#>   3: -0.9080038 -0.6117581 -0.8885315 -0.6456336 -0.9021218 -0.6992034
#>   4: -0.9134948 -0.6192009 -0.8989011 -0.6481388 -0.9077991 -0.7091047
#>   5: -0.9188258 -0.6264750 -0.9012332 -0.6513934 -0.9091613 -0.7153334
#>  ---                                                                  
#> 423: -0.9997684 -0.6440702 -0.9665285 -0.6262602 -0.9431579 -0.6736915
#> 424: -1.0026237 -0.6445356 -0.9648960 -0.6271436 -0.9473616 -0.6741095
#> 425: -1.0109296 -0.6428438 -0.9644467 -0.6285679 -0.9552059 -0.6747956
#> 426: -1.0173761 -0.6405254 -0.9617673 -0.6288394 -0.9559735 -0.6754367
#> 427: -1.0150208 -0.6404064 -0.9559705 -0.6282292 -0.9501902 -0.6762650
#>             1_9       1_10       1_11       1_12       1_13       1_14
#>   1: -0.9222420 -0.7542071 -0.9433687 -0.7402943 -0.9332132 -0.7567033
#>   2: -0.9288093 -0.7519819 -0.9429623 -0.7404103 -0.9356532 -0.7536005
#>   3: -0.9341334 -0.7532891 -0.9469777 -0.7458370 -0.9353976 -0.7510077
#>   4: -0.9392776 -0.7558519 -0.9532304 -0.7527444 -0.9364607 -0.7533079
#>   5: -0.9439399 -0.7587993 -0.9586670 -0.7597492 -0.9411145 -0.7598255
#>  ---                                                                  
#> 423: -0.9860366 -0.7275399 -0.9848989 -0.7450432 -0.9783244 -0.7571870
#> 424: -0.9909912 -0.7274999 -0.9867297 -0.7447987 -0.9787123 -0.7569463
#> 425: -0.9958966 -0.7286643 -0.9885115 -0.7436514 -0.9769683 -0.7576035
#> 426: -0.9968778 -0.7298300 -0.9915698 -0.7418910 -0.9733995 -0.7570600
#> 427: -0.9968597 -0.7295656 -0.9930958 -0.7406996 -0.9704962 -0.7558029
#>            1_15        2_0        2_1        2_2        2_3        2_4
#>   1: -0.8721603 -0.6341125 -0.8133104 -0.3901595 -0.5201151 -0.1980942
#>   2: -0.8671569 -0.6482702 -0.8365884 -0.4006985 -0.5343671 -0.2144724
#>   3: -0.8655049 -0.6606091 -0.8611584 -0.4139356 -0.5486515 -0.2310214
#>   4: -0.8679831 -0.6671489 -0.8794094 -0.4261746 -0.5613412 -0.2464838
#>   5: -0.8755950 -0.6697885 -0.8844213 -0.4350373 -0.5719566 -0.2595440
#>  ---                                                                  
#> 423: -0.8880167 -0.6340107 -0.9397699 -0.4569200 -0.6700193 -0.2245990
#> 424: -0.8865166 -0.6315157 -0.9405787 -0.4589075 -0.6707723 -0.2224570
#> 425: -0.8867059 -0.6310282 -0.9495268 -0.4598360 -0.6808237 -0.2207177
#> 426: -0.8891636 -0.6334830 -0.9544747 -0.4595112 -0.6913007 -0.2204102
#> 427: -0.8905814 -0.6349432 -0.9491858 -0.4599341 -0.6900923 -0.2235211
#>             2_5        2_6        2_7        2_8        2_9       2_10
#>   1: -0.4099849 -0.2527224 -0.4977783 -0.3796127 -0.6079962 -0.5419455
#>   2: -0.4390349 -0.2684393 -0.5227140 -0.3849098 -0.6232019 -0.5477891
#>   3: -0.4656750 -0.2806208 -0.5420383 -0.3974718 -0.6357659 -0.5548076
#>   4: -0.4813786 -0.2890445 -0.5488055 -0.4114504 -0.6433610 -0.5590054
#>   5: -0.4849669 -0.2930531 -0.5470268 -0.4193472 -0.6473718 -0.5607016
#>  ---                                                                  
#> 423: -0.5154879 -0.2002706 -0.5602914 -0.3915415 -0.7324996 -0.5452767
#> 424: -0.5099697 -0.1984237 -0.5632606 -0.3922307 -0.7367614 -0.5450110
#> 425: -0.5091169 -0.1995823 -0.5696034 -0.3931721 -0.7409591 -0.5462835
#> 426: -0.5092536 -0.2019320 -0.5696742 -0.3938283 -0.7417867 -0.5476882
#> 427: -0.5064613 -0.2023355 -0.5630394 -0.3950130 -0.7414424 -0.5475457
#>            2_11       2_12       2_13       2_14       2_15        3_0
#>   1: -0.7610027 -0.6229805 -0.8415435 -0.7059712 -0.8520501 -0.3957456
#>   2: -0.7661793 -0.6255116 -0.8423542 -0.7018267 -0.8484985 -0.4134940
#>   3: -0.7744415 -0.6338600 -0.8423946 -0.6993682 -0.8479570 -0.4314153
#>   4: -0.7820604 -0.6428714 -0.8442754 -0.7030330 -0.8511558 -0.4433241
#>   5: -0.7879329 -0.6515540 -0.8493459 -0.7104161 -0.8589715 -0.4486529
#>  ---                                                                  
#> 423: -0.8237435 -0.6370554 -0.8907275 -0.7200700 -0.8903002 -0.4140871
#> 424: -0.8245081 -0.6377808 -0.8907435 -0.7203900 -0.8880396 -0.4108835
#> 425: -0.8265153 -0.6387181 -0.8892012 -0.7209251 -0.8868613 -0.4105957
#> 426: -0.8301792 -0.6393422 -0.8864406 -0.7207760 -0.8885221 -0.4125072
#> 427: -0.8323184 -0.6391008 -0.8829991 -0.7205515 -0.8889065 -0.4132366
#>             3_1         3_2        3_3          3_4        3_5          3_6
#>   1: -0.4706071 -0.03016547 -0.2407176  0.082791172 -0.2010814  0.065702595
#>   2: -0.5057679 -0.05242888 -0.2619650  0.057928637 -0.2434139  0.022760876
#>   3: -0.5376490 -0.07693648 -0.2832872  0.030258324 -0.2863251 -0.012867784
#>   4: -0.5588946 -0.09690103 -0.3049061  0.005039054 -0.3173652 -0.033973984
#>   5: -0.5629396 -0.10788352 -0.3245930 -0.014849934 -0.3317456 -0.044267047
#>  ---                                                                       
#> 423: -0.6419066 -0.07969962 -0.4375644  0.015281066 -0.3982216  0.018667474
#> 424: -0.6423715 -0.08082324 -0.4397005  0.012953204 -0.3960228  0.017094024
#> 425: -0.6495159 -0.08243242 -0.4492351  0.011830383 -0.3970601  0.011106797
#> 426: -0.6525790 -0.08291471 -0.4578898  0.012850599 -0.3957864  0.005810447
#> 427: -0.6471939 -0.08415753 -0.4561356  0.012156048 -0.3883185  0.006481665
#>             3_7         3_8        3_9       3_10       3_11       3_12
#>   1: -0.1998793 -0.01399566 -0.2785261 -0.1904907 -0.5733052 -0.4793909
#>   2: -0.2449123 -0.03674952 -0.3024972 -0.2085340 -0.5838028 -0.4878016
#>   3: -0.2837758 -0.06542999 -0.3245068 -0.2265603 -0.5959360 -0.5013934
#>   4: -0.3092131 -0.09077036 -0.3431188 -0.2380186 -0.6060463 -0.5143723
#>   5: -0.3230577 -0.10647227 -0.3560015 -0.2436821 -0.6122009 -0.5251434
#>  ---                                                                   
#> 423: -0.3698645 -0.01980718 -0.3975851 -0.1807054 -0.6520342 -0.4950348
#> 424: -0.3754378 -0.02023140 -0.4007568 -0.1797360 -0.6545247 -0.4978326
#> 425: -0.3863423 -0.02436409 -0.4062619 -0.1799946 -0.6574040 -0.4994102
#> 426: -0.3862761 -0.02749851 -0.4110578 -0.1800945 -0.6603731 -0.4983249
#> 427: -0.3753696 -0.02886785 -0.4138399 -0.1785197 -0.6617479 -0.4960453
#>            3_13       3_14       3_15         4_0        4_1        4_2
#>   1: -0.7665755 -0.6297212 -0.8223611 -0.06994937 -0.1948373 0.23507126
#>   2: -0.7730041 -0.6271325 -0.8182341 -0.09679992 -0.2343327 0.19651763
#>   3: -0.7757829 -0.6259195 -0.8187702 -0.12339281 -0.2791846 0.15964492
#>   4: -0.7789082 -0.6308877 -0.8228862 -0.14147647 -0.3141552 0.13575174
#>   5: -0.7851093 -0.6393642 -0.8300613 -0.15012543 -0.3275808 0.12282108
#>  ---                                                                   
#> 423: -0.8091943 -0.6442823 -0.8769802 -0.10772412 -0.4471408 0.03666973
#> 424: -0.8092830 -0.6438907 -0.8746197 -0.10918775 -0.4476721 0.03725535
#> 425: -0.8063261 -0.6447943 -0.8734123 -0.10935027 -0.4572730 0.03887363
#> 426: -0.8015150 -0.6445024 -0.8748469 -0.11035092 -0.4627162 0.03930071
#> 427: -0.7982160 -0.6434273 -0.8757001 -0.11206047 -0.4563209 0.03814443
#>              4_3        4_4         4_5        4_6         4_7        4_8
#>   1: -0.02331112 0.26986513 -0.03277086 0.19152664 -0.06653646 0.12510364
#>   2: -0.06179452 0.23874061 -0.08068238 0.15103769 -0.10559224 0.10004297
#>   3: -0.10245890 0.20027897 -0.12478228 0.11091606 -0.15335120 0.06658314
#>   4: -0.13633749 0.16619243 -0.15494780 0.08322451 -0.18932059 0.03945834
#>   5: -0.15830097 0.14300099 -0.16945702 0.06948098 -0.20651995 0.02460175
#>  ---                                                                     
#> 423: -0.39198825 0.05423059 -0.37294900 0.05756936 -0.32915789 0.07003556
#> 424: -0.39235875 0.05655045 -0.36823896 0.06076564 -0.33180121 0.07390746
#> 425: -0.39971539 0.05786621 -0.36747903 0.05863204 -0.33813080 0.07251260
#> 426: -0.40848437 0.05763303 -0.36740896 0.05429918 -0.33687240 0.06923430
#> 427: -0.41004351 0.05724512 -0.36362541 0.05428709 -0.32945213 0.06637439
#>             4_9         4_10       4_11       4_12       4_13       4_14
#>   1: -0.1101248  0.029535320 -0.3481123 -0.3506066 -0.6812164 -0.5529566
#>   2: -0.1527066  0.003534036 -0.3676054 -0.3633745 -0.6913838 -0.5519550
#>   3: -0.1881872 -0.025381880 -0.3941703 -0.3797562 -0.6990299 -0.5533080
#>   4: -0.2133017 -0.047853392 -0.4153597 -0.3951810 -0.7052442 -0.5597531
#>   5: -0.2292614 -0.060071301 -0.4268375 -0.4070649 -0.7121986 -0.5681785
#>  ---                                                                    
#> 423: -0.2929688  0.017835222 -0.4555726 -0.3775359 -0.7328435 -0.5438783
#> 424: -0.2945867  0.019689614 -0.4546731 -0.3758536 -0.7310277 -0.5421190
#> 425: -0.2965241  0.016919773 -0.4559585 -0.3765136 -0.7278451 -0.5422699
#> 426: -0.2965332  0.012936818 -0.4604053 -0.3785573 -0.7247259 -0.5428804
#> 427: -0.2969316  0.011210520 -0.4643245 -0.3780981 -0.7230524 -0.5427254
#>            4_15        5_0         5_1        5_2         5_3        5_4
#>   1: -0.7702336 0.11460015 -0.03510817 0.34293243  0.11290687 0.42655033
#>   2: -0.7664602 0.08463844 -0.08284651 0.31294748  0.06828209 0.38676611
#>   3: -0.7678913 0.05434874 -0.13940026 0.27856693  0.01997435 0.33720395
#>   4: -0.7723041 0.03276557 -0.18394655 0.25130591 -0.01962078 0.29737791
#>   5: -0.7798773 0.02154485 -0.20320033 0.23551796 -0.04708247 0.27245685
#>  ---                                                                    
#> 423: -0.8315062 0.03843024 -0.37421671 0.07204729 -0.36881086 0.07357048
#> 424: -0.8301061 0.03988123 -0.37570393 0.07378091 -0.36943567 0.07481265
#> 425: -0.8283818 0.03883109 -0.38774556 0.07472437 -0.38019809 0.07525086
#> 426: -0.8283017 0.03590450 -0.39282519 0.07751127 -0.39020240 0.07412328
#> 427: -0.8286900 0.03426958 -0.38598010 0.07995838 -0.38810304 0.07064135
#>              5_5        5_6          5_7        5_8         5_9       5_10
#>   1:  0.11753017 0.32303247  0.109790646 0.26948279  0.01834719 0.12938540
#>   2:  0.06776483 0.28515229  0.058814906 0.23558564 -0.02757971 0.09512787
#>   3:  0.01658157 0.24822997  0.007717322 0.20105782 -0.06737868 0.06089470
#>   4: -0.02130803 0.22373028 -0.032182094 0.17547110 -0.09359654 0.03801979
#>   5: -0.04069167 0.21401943 -0.059577677 0.16066670 -0.10749152 0.02652283
#>  ---                                                                      
#> 423: -0.35792434 0.07230426 -0.308467656 0.08268414 -0.26994801 0.05796900
#> 424: -0.35649100 0.07285604 -0.310336679 0.08556347 -0.27288014 0.06006217
#> 425: -0.35849795 0.07272252 -0.320685357 0.08541255 -0.27874035 0.05942376
#> 426: -0.35917634 0.07051150 -0.325145066 0.08421563 -0.28221530 0.05715792
#> 427: -0.35430756 0.06807132 -0.319326639 0.08278632 -0.28278291 0.05654842
#>            5_11       5_12       5_13       5_14       5_15        6_0
#>   1: -0.1717014 -0.1969883 -0.5527974 -0.4865161 -0.7210248 0.18089953
#>   2: -0.2030258 -0.2211370 -0.5673119 -0.4884878 -0.7198143 0.15293343
#>   3: -0.2327487 -0.2465808 -0.5776873 -0.4922220 -0.7229511 0.12460253
#>   4: -0.2546774 -0.2635905 -0.5845709 -0.4998491 -0.7282363 0.10344046
#>   5: -0.2675154 -0.2707982 -0.5896380 -0.5077798 -0.7358948 0.08880223
#>  ---                                                                  
#> 423: -0.2607206 -0.1880689 -0.6150235 -0.4683118 -0.7755380 0.13000481
#> 424: -0.2616906 -0.1876739 -0.6141766 -0.4669085 -0.7738265 0.13281861
#> 425: -0.2635156 -0.1880164 -0.6130396 -0.4668194 -0.7734500 0.13485433
#> 426: -0.2661669 -0.1878610 -0.6108111 -0.4666386 -0.7754002 0.13375901
#> 427: -0.2657728 -0.1861850 -0.6090330 -0.4672388 -0.7765176 0.13073599
#>              6_1       6_2         6_3        6_4         6_5        6_6
#>   1: -0.01905255 0.4086876  0.11888564 0.48320299  0.16874145 0.44582400
#>   2: -0.05683843 0.3770023  0.08798480 0.44471416  0.12288575 0.41802159
#>   3: -0.10118315 0.3361670  0.05229931 0.39902419  0.07216766 0.38312700
#>   4: -0.13947496 0.3031361  0.01808203 0.35593370  0.03447116 0.35206458
#>   5: -0.15685874 0.2850990 -0.01346898 0.32451251  0.01535929 0.32730293
#>  ---                                                                    
#> 423: -0.34831402 0.1095677 -0.34016314 0.09610651 -0.34881091 0.08453228
#> 424: -0.34806740 0.1117842 -0.34011540 0.09830205 -0.34332514 0.08498331
#> 425: -0.35857370 0.1093288 -0.34659386 0.09833130 -0.34261155 0.08149479
#> 426: -0.36376283 0.1050092 -0.35318628 0.09690438 -0.34267443 0.07795043
#> 427: -0.35765105 0.1034839 -0.35265785 0.09591320 -0.33919519 0.07804901
#>              6_7        6_8         6_9       6_10         6_11         6_12
#>   1:  0.21916731 0.43569326  0.19325805 0.28649321  0.021155749  0.048674889
#>   2:  0.16726121 0.41328928  0.14087676 0.25508243 -0.009800109  0.024229340
#>   3:  0.11559406 0.37692678  0.10097029 0.22570246 -0.042084817 -0.007773926
#>   4:  0.07695936 0.34050414  0.07249004 0.20341654 -0.067647710 -0.031864427
#>   5:  0.05130208 0.31644794  0.05302528 0.18978401 -0.082216792 -0.040720206
#>  ---                                                                        
#> 423: -0.30920589 0.09023030 -0.26208341 0.07535793 -0.207132086  0.040714394
#> 424: -0.31480187 0.08940946 -0.26722503 0.07533854 -0.203569978  0.040153250
#> 425: -0.32193980 0.08624943 -0.27583653 0.07458680 -0.207171693  0.038982097
#> 426: -0.31959346 0.08395305 -0.27948758 0.07265072 -0.215214461  0.038374390
#> 427: -0.31122905 0.08047207 -0.27833202 0.07338902 -0.218296513  0.039154116
#>            6_13       6_14       6_15        7_0         7_1       7_2
#>   1: -0.2817131 -0.3330086 -0.6200074 0.17514743 -0.09557199 0.4060850
#>   2: -0.3051751 -0.3411041 -0.6209924 0.14219251 -0.12218226 0.3676866
#>   3: -0.3258636 -0.3490969 -0.6252375 0.11172757 -0.15664205 0.3342102
#>   4: -0.3378293 -0.3588005 -0.6312293 0.09271412 -0.18458751 0.3104898
#>   5: -0.3440845 -0.3673511 -0.6388685 0.08287866 -0.19328009 0.2957985
#>  ---                                                                  
#> 423: -0.3216147 -0.3059025 -0.6783990 0.14231284 -0.33671686 0.1486460
#> 424: -0.3206956 -0.3035870 -0.6754252 0.14542726 -0.33683294 0.1504826
#> 425: -0.3180229 -0.3014976 -0.6750098 0.14393683 -0.34644127 0.1512316
#> 426: -0.3146592 -0.3007114 -0.6782953 0.14145134 -0.35066342 0.1511961
#> 427: -0.3124145 -0.3003827 -0.6798649 0.14077842 -0.34340900 0.1495800
#>              7_3       7_4         7_5       7_6         7_7        7_8
#>   1:  0.10227323 0.4955553  0.16690482 0.4960505  0.22606452 0.51621252
#>   2:  0.06227709 0.4627217  0.11452152 0.4559644  0.19545102 0.47442657
#>   3:  0.02216644 0.4177838  0.07430249 0.4193894  0.15471040 0.42696881
#>   4: -0.01016116 0.3727280  0.04814187 0.3901626  0.11800983 0.39333984
#>   5: -0.02548500 0.3412454  0.03500795 0.3708753  0.09266277 0.37308314
#>  ---                                                                   
#> 423: -0.31263500 0.1300725 -0.32309648 0.1092520 -0.30024305 0.10954092
#> 424: -0.31076008 0.1306683 -0.31835282 0.1115749 -0.30168831 0.10853294
#> 425: -0.31428114 0.1317825 -0.31976902 0.1091268 -0.30967656 0.10527629
#> 426: -0.32064453 0.1312346 -0.32075778 0.1039964 -0.31268883 0.10104267
#> 427: -0.32244471 0.1287172 -0.31694186 0.1019832 -0.30719063 0.09788156
#>             7_9       7_10        7_11      7_12       7_13       7_14
#>   1:  0.2669946 0.43942773  0.17619297 0.2079387 -0.1004790 -0.1541892
#>   2:  0.2206271 0.40604094  0.15137148 0.1752643 -0.1278359 -0.1677137
#>   3:  0.1714102 0.36487803  0.11816086 0.1402042 -0.1462014 -0.1811404
#>   4:  0.1325444 0.32844305  0.08585792 0.1160336 -0.1566433 -0.1942294
#>   5:  0.1089885 0.30745614  0.06451309 0.1055136 -0.1644206 -0.2030196
#>  ---                                                                  
#> 423: -0.2565538 0.10326929 -0.18363880 0.1079208 -0.1985933 -0.1611951
#> 424: -0.2593665 0.10223637 -0.18624392 0.1098215 -0.1970126 -0.1608094
#> 425: -0.2634064 0.10008022 -0.18872470 0.1085010 -0.1931469 -0.1615037
#> 426: -0.2673599 0.09678093 -0.19275670 0.1058363 -0.1889191 -0.1614061
#> 427: -0.2727987 0.09452595 -0.19443908 0.1043969 -0.1866418 -0.1601589
#>            7_15          8_0        8_1       8_2         8_3       8_4
#>   1: -0.5388833  0.087934248 -0.1846190 0.3153276 -0.02330378 0.4506199
#>   2: -0.5411273  0.071245745 -0.2132872 0.2905726 -0.05313157 0.4143729
#>   3: -0.5462263  0.051149037 -0.2485526 0.2634043 -0.08097915 0.3769518
#>   4: -0.5518396  0.036638767 -0.2723057 0.2415955 -0.10360466 0.3495513
#>   5: -0.5576556  0.032449901 -0.2760711 0.2275214 -0.11793862 0.3316274
#>  ---                                                                   
#> 423: -0.5895592  0.003167340 -0.3458221 0.1443959 -0.28520176 0.1748143
#> 424: -0.5865974  0.006487147 -0.3453450 0.1424087 -0.28314653 0.1798537
#> 425: -0.5859556  0.006562222 -0.3560074 0.1408821 -0.29136106 0.1807242
#> 426: -0.5888314  0.002249188 -0.3626761 0.1408205 -0.29821774 0.1776260
#> 427: -0.5904717 -0.003012084 -0.3560008 0.1408087 -0.29326501 0.1746140
#>              8_5       8_6         8_7       8_8        8_9      8_10
#>   1:  0.15336494 0.5026962  0.25929761 0.5448662  0.2914203 0.4914328
#>   2:  0.10891031 0.4816248  0.21767247 0.5082873  0.2498384 0.4499103
#>   3:  0.06503499 0.4589221  0.16304955 0.4671142  0.1989192 0.4126056
#>   4:  0.03334329 0.4330226  0.11946984 0.4280581  0.1616991 0.3867722
#>   5:  0.01523896 0.4074669  0.09858868 0.4038964  0.1397330 0.3701818
#>  ---                                                                 
#> 423: -0.28495014 0.1534336 -0.25919485 0.1541284 -0.2198540 0.1424541
#> 424: -0.28366613 0.1540914 -0.25829354 0.1559501 -0.2187341 0.1477413
#> 425: -0.28505880 0.1516479 -0.26450428 0.1500032 -0.2210669 0.1490496
#> 426: -0.28465354 0.1485623 -0.26641917 0.1438126 -0.2242813 0.1446446
#> 427: -0.27912781 0.1472479 -0.26357475 0.1420557 -0.2270529 0.1418104
#>            8_11      8_12        8_13       8_14       8_15        9_0
#>   1:  0.2251298 0.3129992 -0.01930906 -0.1302575 -0.4997482 -0.1404191
#>   2:  0.1958722 0.2850758 -0.04237487 -0.1349224 -0.5037256 -0.1482290
#>   3:  0.1697999 0.2484965 -0.05983021 -0.1419368 -0.5089787 -0.1594115
#>   4:  0.1452528 0.2202435 -0.07422929 -0.1524901 -0.5153573 -0.1682616
#>   5:  0.1240693 0.2049263 -0.08699963 -0.1633783 -0.5238560 -0.1715885
#>  ---                                                                  
#> 423: -0.1471200 0.1452986 -0.20291974 -0.2383699 -0.5793382 -0.2771131
#> 424: -0.1480197 0.1500701 -0.20409964 -0.2370797 -0.5767770 -0.2751649
#> 425: -0.1519214 0.1520038 -0.20630156 -0.2374736 -0.5755845 -0.2740324
#> 426: -0.1592668 0.1524178 -0.20555831 -0.2382059 -0.5769823 -0.2760093
#> 427: -0.1644321 0.1516853 -0.20182349 -0.2386301 -0.5775211 -0.2802055
#>             9_1       9_2        9_3       9_4         9_5       9_6
#>   1: -0.3880415 0.1964528 -0.1848320 0.3294951  0.06159078 0.5099736
#>   2: -0.4016658 0.1757130 -0.1974400 0.3060773  0.01614512 0.4669406
#>   3: -0.4255078 0.1511883 -0.2236059 0.2722496 -0.02511934 0.4293555
#>   4: -0.4447283 0.1303705 -0.2497362 0.2399727 -0.05653130 0.3961539
#>   5: -0.4463654 0.1197424 -0.2649184 0.2176039 -0.07012609 0.3755288
#>  ---                                                                
#> 423: -0.4726446 0.1334595 -0.2893841 0.2214528 -0.23276733 0.2040207
#> 424: -0.4710627 0.1328289 -0.2894299 0.2196637 -0.22747667 0.2037956
#> 425: -0.4767421 0.1337296 -0.2961083 0.2174471 -0.23046409 0.2013982
#> 426: -0.4798250 0.1356475 -0.3010665 0.2144648 -0.23281474 0.1966529
#> 427: -0.4757109 0.1343473 -0.2967418 0.2124132 -0.22850421 0.1938415
#>              9_7       9_8        9_9      9_10       9_11      9_12
#>   1:  0.17453891 0.5510354  0.2190386 0.4883946  0.2343115 0.3436384
#>   2:  0.13835642 0.5170754  0.2014254 0.4554164  0.1979440 0.3181358
#>   3:  0.10251222 0.4706872  0.1717279 0.4163522  0.1590499 0.2883997
#>   4:  0.07783717 0.4303708  0.1354595 0.3866921  0.1322260 0.2594654
#>   5:  0.06698119 0.4054990  0.1095207 0.3700014  0.1164159 0.2401658
#>  ---                                                                
#> 423: -0.20711856 0.1984591 -0.1697335 0.1919664 -0.1002033 0.1958435
#> 424: -0.20930032 0.1975257 -0.1713480 0.1919565 -0.1014204 0.1963937
#> 425: -0.21713974 0.1973275 -0.1762835 0.1880476 -0.1064350 0.1965378
#> 426: -0.21859702 0.1965215 -0.1810040 0.1842630 -0.1137590 0.1959506
#> 427: -0.21456438 0.1942090 -0.1817123 0.1841256 -0.1176966 0.1958615
#>             9_13       9_14       9_15       10_0       10_1         10_2
#>   1: -0.03961723 -0.2325946 -0.5138462 -0.3454294 -0.5610783 -0.003405581
#>   2: -0.06623416 -0.2391484 -0.5182777 -0.3505079 -0.5801501 -0.016585065
#>   3: -0.08755251 -0.2452990 -0.5224738 -0.3595735 -0.6064633 -0.035956357
#>   4: -0.10419883 -0.2539910 -0.5295882 -0.3662870 -0.6256430 -0.051071431
#>   5: -0.11806381 -0.2648067 -0.5404692 -0.3689005 -0.6284144 -0.058754291
#>  ---                                                                     
#> 423: -0.20896144 -0.3554857 -0.5956661 -0.4831747 -0.6749610 -0.006982267
#> 424: -0.20708182 -0.3550624 -0.5945614 -0.4809514 -0.6780014 -0.005729941
#> 425: -0.20645618 -0.3566869 -0.5941038 -0.4795221 -0.6867968 -0.004676747
#> 426: -0.20509402 -0.3572714 -0.5957916 -0.4812161 -0.6908116 -0.004938770
#> 427: -0.20455551 -0.3560388 -0.5966506 -0.4826814 -0.6844195 -0.006340539
#>            10_3       10_4        10_5      10_6         10_7      10_8
#>   1: -0.3789502 0.13877778 -0.09041325 0.3947443  0.096415646 0.5012836
#>   2: -0.3927487 0.11653356 -0.12458400 0.3548783  0.067286357 0.4652751
#>   3: -0.4124519 0.08843436 -0.15803877 0.3201760  0.033194412 0.4181609
#>   4: -0.4331997 0.06170695 -0.18525149 0.2936058  0.004564124 0.3783152
#>   5: -0.4479685 0.04130512 -0.20156626 0.2748121 -0.016045203 0.3591099
#>  ---                                                                   
#> 423: -0.4192245 0.21815482 -0.19248809 0.2559083 -0.168943003 0.2504518
#> 424: -0.4255278 0.22156312 -0.19656545 0.2575580 -0.171470091 0.2539417
#> 425: -0.4362897 0.22243653 -0.20338295 0.2567109 -0.178340837 0.2579956
#> 426: -0.4440649 0.22056223 -0.20394890 0.2551845 -0.180030599 0.2567944
#> 427: -0.4405728 0.21886383 -0.19539179 0.2556418 -0.175978899 0.2496768
#>             10_9     10_10       10_11     10_12       10_13      10_14
#>   1:  0.16088957 0.4022605  0.12462527 0.3303100 -0.07084840 -0.3302290
#>   2:  0.13614024 0.3802870  0.10487096 0.3162005 -0.08289412 -0.3306039
#>   3:  0.10655165 0.3530020  0.08353272 0.2910916 -0.09676377 -0.3353746
#>   4:  0.08416789 0.3325290  0.06740724 0.2647256 -0.10807587 -0.3462425
#>   5:  0.06685340 0.3216399  0.05747279 0.2419084 -0.12040044 -0.3579053
#>  ---                                                                   
#> 423: -0.11701809 0.2348692 -0.06368053 0.2281845 -0.15948832 -0.4201501
#> 424: -0.12409801 0.2383129 -0.06182728 0.2300553 -0.15933338 -0.4195252
#> 425: -0.13336122 0.2388805 -0.06120444 0.2304868 -0.15755838 -0.4191516
#> 426: -0.13745880 0.2341105 -0.06617249 0.2291671 -0.15389903 -0.4183375
#> 427: -0.13737124 0.2278370 -0.07184657 0.2277900 -0.15289837 -0.4172875
#>           10_15       11_0       11_1       11_2       11_3        11_4
#>   1: -0.5502016 -0.4089794 -0.7035590 -0.1988441 -0.6013403 -0.04612614
#>   2: -0.5491410 -0.4128581 -0.7133205 -0.2056266 -0.6083025 -0.06301054
#>   3: -0.5519719 -0.4174158 -0.7311138 -0.2184241 -0.6193420 -0.08642311
#>   4: -0.5582104 -0.4215689 -0.7447819 -0.2302519 -0.6318524 -0.10882086
#>   5: -0.5680800 -0.4243726 -0.7436408 -0.2369744 -0.6402031 -0.12603880
#>  ---                                                                   
#> 423: -0.6367587 -0.5125267 -0.8724214 -0.3014192 -0.6272018  0.09982392
#> 424: -0.6344557 -0.5105906 -0.8761734 -0.3029543 -0.6285866  0.10479824
#> 425: -0.6346116 -0.5097994 -0.8845704 -0.3034330 -0.6365673  0.10503902
#> 426: -0.6366958 -0.5112385 -0.8873740 -0.3033138 -0.6447878  0.10215051
#> 427: -0.6373948 -0.5125880 -0.8805217 -0.3047109 -0.6447583  0.09967330
#>            11_5      11_6        11_7      11_8         11_9     11_10
#>   1: -0.3021611 0.1909627 -0.06505151 0.3536376  0.001076699 0.2811520
#>   2: -0.3240646 0.1635235 -0.09604929 0.3317861 -0.020703882 0.2534717
#>   3: -0.3507667 0.1393459 -0.12998015 0.3020550 -0.043339010 0.2297511
#>   4: -0.3766404 0.1217387 -0.15413399 0.2728064 -0.061574921 0.2129553
#>   5: -0.3931606 0.1105335 -0.16765058 0.2570367 -0.073306367 0.2012097
#>  ---                                                                  
#> 423: -0.2195135 0.2539870 -0.14811064 0.2794592 -0.083007313 0.2626795
#> 424: -0.2189381 0.2545430 -0.15172820 0.2852612 -0.081238411 0.2603240
#> 425: -0.2219534 0.2513329 -0.15867922 0.2861450 -0.083784021 0.2556774
#> 426: -0.2223453 0.2494126 -0.15801638 0.2827066 -0.087282449 0.2529730
#> 427: -0.2186473 0.2508664 -0.15099348 0.2778258 -0.093390450 0.2562906
#>            11_11     11_12      11_13      11_14      11_15       12_0
#>   1: -0.04550443 0.1613120 -0.2492418 -0.3870185 -0.5538554 -0.5889121
#>   2: -0.05919959 0.1532610 -0.2606162 -0.3864829 -0.5549448 -0.5879138
#>   3: -0.07607881 0.1355628 -0.2675352 -0.3900220 -0.5566438 -0.5898755
#>   4: -0.09350286 0.1189713 -0.2733023 -0.4005955 -0.5599077 -0.5909628
#>   5: -0.10829892 0.1061453 -0.2821340 -0.4124226 -0.5665730 -0.5905932
#>  ---                                                                  
#> 423: -0.04777495 0.1684111 -0.2748755 -0.4656470 -0.6215362 -0.6319057
#> 424: -0.04765457 0.1688343 -0.2724172 -0.4665069 -0.6193126 -0.6320986
#> 425: -0.04815739 0.1672493 -0.2692471 -0.4676030 -0.6189794 -0.6328617
#> 426: -0.05001089 0.1658563 -0.2664106 -0.4671336 -0.6216958 -0.6342685
#> 427: -0.05160535 0.1668420 -0.2649757 -0.4658949 -0.6232929 -0.6346731
#>            12_1       12_2       12_3       12_4       12_5         12_6
#>   1: -0.9084547 -0.4400639 -0.7794662 -0.2471266 -0.5029922  0.001548128
#>   2: -0.9101344 -0.4375786 -0.7770191 -0.2589712 -0.5270271 -0.022970274
#>   3: -0.9202237 -0.4420333 -0.7809778 -0.2757573 -0.5534080 -0.042182062
#>   4: -0.9293817 -0.4489326 -0.7892566 -0.2921411 -0.5725283 -0.055214901
#>   5: -0.9286409 -0.4542748 -0.7958061 -0.3046873 -0.5792183 -0.064597100
#>  ---                                                                    
#> 423: -1.0792398 -0.5509068 -0.7931226 -0.1663056 -0.3932111  0.167518958
#> 424: -1.0811239 -0.5530722 -0.7953201 -0.1666212 -0.3933381  0.164244816
#> 425: -1.0896326 -0.5551605 -0.8030046 -0.1681725 -0.3954019  0.159555063
#> 426: -1.0944729 -0.5558071 -0.8086336 -0.1697897 -0.3963541  0.157658920
#> 427: -1.0899518 -0.5560302 -0.8055803 -0.1712147 -0.3940453  0.159509078
#>            12_7       12_8       12_9        12_10      12_11      12_12
#>   1: -0.2636845 0.15986118 -0.2198299  0.012900840 -0.3475010 -0.1755732
#>   2: -0.2925263 0.13392644 -0.2365713  0.002008484 -0.3510732 -0.1797495
#>   3: -0.3226509 0.10030603 -0.2550317 -0.013462396 -0.3556432 -0.1896459
#>   4: -0.3424500 0.06911202 -0.2684755 -0.026638059 -0.3618752 -0.1994632
#>   5: -0.3514164 0.04984137 -0.2740572 -0.033214808 -0.3696662 -0.2079690
#>  ---                                                                    
#> 423: -0.2231302 0.23791830 -0.1277080  0.222949252 -0.3009740 -0.2151229
#> 424: -0.2291479 0.23654796 -0.1266753  0.221743941 -0.3037770 -0.2165980
#> 425: -0.2400716 0.23113360 -0.1276943  0.217651576 -0.3070489 -0.2181160
#> 426: -0.2403418 0.22415788 -0.1321646  0.217663661 -0.3105931 -0.2197210
#> 427: -0.2322121 0.22267960 -0.1382107  0.219312802 -0.3126730 -0.2208407
#>           12_13      12_14      12_15
#>   1: -0.5086929 -0.4955095 -0.6490921
#>   2: -0.5129193 -0.4911957 -0.6464767
#>   3: -0.5125197 -0.4882756 -0.6439207
#>   4: -0.5119967 -0.4921051 -0.6446184
#>   5: -0.5161167 -0.4999369 -0.6508147
#>  ---                                 
#> 423: -0.5514245 -0.5360808 -0.7430303
#> 424: -0.5530964 -0.5362788 -0.7415859
#> 425: -0.5516642 -0.5378205 -0.7405182
#> 426: -0.5466566 -0.5389180 -0.7409301
#> 427: -0.5412602 -0.5395846 -0.7407396
#> 
#> $metadata
#>       x  y       file_name  license
#>   1:  0  0 CA small UF.dat CC BY-NC
#>   2:  0  1 CA small UF.dat CC BY-NC
#>   3:  0  2 CA small UF.dat CC BY-NC
#>   4:  0  3 CA small UF.dat CC BY-NC
#>   5:  0  4 CA small UF.dat CC BY-NC
#>  ---                               
#> 204: 12 11 CA small UF.dat CC BY-NC
#> 205: 12 12 CA small UF.dat CC BY-NC
#> 206: 12 13 CA small UF.dat CC BY-NC
#> 207: 12 14 CA small UF.dat CC BY-NC
#> 208: 12 15 CA small UF.dat CC BY-NC
#>                                                       description samples lines
#>   1: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>   2: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>   3: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>   4: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>   5: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>  ---                                                                           
#> 204: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 205: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 206: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 207: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#> 208: Pixel 15, XPos=12549.72, YPos= 589.82, X=12362.22, Y= 614.82      16    13
#>      bands header offset data type interleave                  z plot titles
#>   1:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>   2:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>   3:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>   4:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>   5:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>  ---                                                                        
#> 204:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 205:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 206:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 207:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#> 208:   427             0         4        bip Wavenumbers (cm-1), Absorbance
#>                    pixel size
#>   1: 0.000025000, 0.000025000
#>   2: 0.000025000, 0.000025000
#>   3: 0.000025000, 0.000025000
#>   4: 0.000025000, 0.000025000
#>   5: 0.000025000, 0.000025000
#>  ---                         
#> 204: 0.000025000, 0.000025000
#> 205: 0.000025000, 0.000025000
#> 206: 0.000025000, 0.000025000
#> 207: 0.000025000, 0.000025000
#> 208: 0.000025000, 0.000025000
#>                                                             session_id
#>   1: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>   2: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>   3: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>   4: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>   5: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>  ---                                                                  
#> 204: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 205: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 206: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 207: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#> 208: d68b222c761bc52ad027c9e82728d82d/8db6004d72093f21d330265ecdc15274
#>                               file_id col_id
#>   1: 215a0fd7e0dc3b1001799214618ce6eb    0_0
#>   2: 215a0fd7e0dc3b1001799214618ce6eb    0_1
#>   3: 215a0fd7e0dc3b1001799214618ce6eb    0_2
#>   4: 215a0fd7e0dc3b1001799214618ce6eb    0_3
#>   5: 215a0fd7e0dc3b1001799214618ce6eb    0_4
#>  ---                                        
#> 204: 215a0fd7e0dc3b1001799214618ce6eb  12_11
#> 205: 215a0fd7e0dc3b1001799214618ce6eb  12_12
#> 206: 215a0fd7e0dc3b1001799214618ce6eb  12_13
#> 207: 215a0fd7e0dc3b1001799214618ce6eb  12_14
#> 208: 215a0fd7e0dc3b1001799214618ce6eb  12_15