Smooth spectral intensities

This smoother can enhance the signal to noise ratio of the data and uses a Savitzky-Golay filter with a running window of data points and the polynomial specified.

Usage

smooth_intens(x, ...)

# S3 method for formula
smooth_intens(formula, data = NULL, ...)

# S3 method for data.frame
smooth_intens(x, ...)

# S3 method for default
smooth_intens(x, y, p = 3, n = 11, make_rel = TRUE, ...)

Arguments

x

a numeric vector containing the spectral wavenumbers; alternatively a data frame containing spectral data as "wavenumber" and "intensity" can be supplied.

formula

an object of class 'formula' of the form intensity ~ wavenumber.

data

a data frame containing the variables in formula.

y

a numeric vector containing the spectral intensities.

p

polynomial order for the filter

n

number of data points in the window, filter length (must be odd).

make_rel

logical; if TRUE spectra are automatically normalized with make_rel().

...

further arguments passed to sgolay().

Value

smooth_intens() returns a data frame containing two columns named "wavenumber" and "intensity".

Details

This is a wrapper around the filter function in the signal package to improve integration with other Open Specy functions. A typical good smooth can be achieved with 11 data point window and a 3rd or 4th order polynomial.

References

Savitzky A, Golay MJ (1964). “Smoothing and Differentiation of Data by Simplified Least Squares Procedures.” Analytical Chemistry, 36(8), 1627--1639.

See also

Author

Win Cowger, Zacharias Steinmetz

Examples

data("raman_hdpe")

smooth_intens(raman_hdpe)
#>     wavenumber  intensity
#> 1      301.040 0.00000000
#> 2      304.632 0.01568318
#> 3      308.221 0.02461353
#> 4      311.810 0.02828915
#> 5      315.398 0.02820811
#> 6      318.983 0.02586852
#> 7      322.566 0.02276014
#> 8      326.150 0.02295130
#> 9      329.732 0.02357879
#> 10     333.311 0.02537400
#> 11     336.889 0.02677859
#> 12     340.467 0.02787982
#> 13     344.042 0.03051446
#> 14     347.618 0.03229305
#> 15     351.190 0.03176529
#> 16     354.760 0.02910156
#> 17     358.332 0.02997839
#> 18     361.900 0.03106715
#> 19     365.466 0.03368102
#> 20     369.031 0.03422955
#> 21     372.597 0.03635306
#> 22     376.160 0.03637799
#> 23     379.721 0.03819814
#> 24     383.280 0.03430851
#> 25     386.839 0.03060588
#> 26     390.396 0.02973321
#> 27     393.951 0.03088846
#> 28     397.505 0.03555934
#> 29     401.058 0.03896692
#> 30     404.610 0.03724651
#> 31     408.160 0.03933261
#> 32     411.708 0.03831034
#> 33     415.256 0.03684757
#> 34     418.801 0.03630735
#> 35     422.346 0.04041722
#> 36     425.887 0.04897773
#> 37     429.429 0.05461686
#> 38     432.970 0.06519282
#> 39     436.507 0.06796044
#> 40     440.044 0.06915309
#> 41     443.581 0.07185838
#> 42     447.116 0.06781915
#> 43     450.647 0.06173122
#> 44     454.180 0.05898022
#> 45     457.708 0.05581366
#> 46     461.238 0.05637882
#> 47     464.764 0.05230635
#> 48     468.291 0.04987533
#> 49     471.815 0.04335522
#> 50     475.339 0.04168883
#> 51     478.859 0.04311004
#> 52     482.379 0.03772440
#> 53     485.899 0.03583361
#> 54     489.416 0.03642786
#> 55     492.931 0.03814827
#> 56     496.447 0.03765376
#> 57     499.960 0.03490691
#> 58     503.471 0.03501912
#> 59     506.981 0.03512716
#> 60     510.489 0.03653590
#> 61     513.999 0.03629904
#> 62     517.503 0.03287899
#> 63     521.009 0.03221825
#> 64     524.510 0.03423787
#> 65     528.014 0.03572557
#> 66     531.513 0.03650266
#> 67     535.014 0.03463680
#> 68     538.510 0.03717586
#> 69     542.008 0.03972324
#> 70     545.502 0.04259059
#> 71     548.997 0.04308511
#> 72     552.488 0.04549950
#> 73     555.979 0.04415309
#> 74     559.469 0.04577793
#> 75     562.955 0.04860788
#> 76     566.442 0.04882397
#> 77     569.928 0.04411154
#> 78     573.411 0.04153923
#> 79     576.894 0.04392453
#> 80     580.376 0.04774352
#> 81     583.854 0.05046543
#> 82     587.331 0.05168301
#> 83     590.811 0.05274269
#> 84     594.286 0.05841922
#> 85     597.759 0.06406250
#> 86     601.231 0.06984707
#> 87     604.705 0.07241938
#> 88     608.174 0.07238614
#> 89     611.642 0.07553607
#> 90     615.109 0.07403590
#> 91     618.574 0.07081533
#> 92     622.038 0.06444897
#> 93     625.503 0.05561835
#> 94     628.966 0.04684176
#> 95     632.426 0.04092005
#> 96     635.884 0.04125665
#> 97     639.342 0.03767038
#> 98     642.798 0.03655668
#> 99     646.252 0.04097822
#> 100    649.706 0.04203790
#> 101    653.158 0.04361702
#> 102    656.608 0.03994764
#> 103    660.058 0.03560921
#> 104    663.506 0.03350233
#> 105    666.953 0.03367686
#> 106    670.398 0.04175947
#> 107    673.842 0.04260721
#> 108    677.285 0.04837932
#> 109    680.727 0.05419299
#> 110    684.165 0.05311669
#> 111    687.602 0.05354471
#> 112    691.040 0.04869099
#> 113    694.476 0.04753158
#> 114    697.911 0.04324302
#> 115    701.345 0.04367936
#> 116    704.775 0.04827959
#> 117    708.204 0.04415725
#> 118    711.634 0.04710771
#> 119    715.063 0.04736120
#> 120    718.490 0.04630568
#> 121    721.912 0.04573221
#> 122    725.336 0.04391207
#> 123    728.760 0.04395778
#> 124    732.180 0.04280253
#> 125    735.599 0.04302693
#> 126    739.018 0.04186752
#> 127    742.434 0.04209192
#> 128    745.849 0.04591922
#> 129    749.265 0.04445229
#> 130    752.675 0.04369182
#> 131    756.088 0.04210854
#> 132    759.498 0.04537899
#> 133    762.906 0.04824219
#> 134    766.313 0.05171626
#> 135    769.719 0.05214013
#> 136    773.124 0.05114694
#> 137    776.527 0.05441323
#> 138    779.929 0.05427610
#> 139    783.330 0.05296709
#> 140    786.729 0.05338680
#> 141    790.127 0.05083112
#> 142    793.524 0.05377327
#> 143    796.919 0.05408078
#> 144    800.312 0.05584691
#> 145    803.705 0.05503241
#> 146    807.097 0.05044049
#> 147    810.485 0.04865775
#> 148    813.874 0.04585688
#> 149    817.262 0.04674202
#> 150    820.647 0.04371676
#> 151    824.032 0.04210439
#> 152    827.416 0.04676280
#> 153    830.797 0.04666722
#> 154    834.179 0.04894033
#> 155    837.557 0.05034907
#> 156    840.936 0.04785572
#> 157    844.312 0.05000000
#> 158    847.688 0.05442985
#> 159    851.061 0.05492021
#> 160    854.433 0.05577211
#> 161    857.806 0.05848155
#> 162    861.175 0.05834441
#> 163    864.543 0.05420130
#> 164    867.910 0.04830868
#> 165    871.278 0.04418218
#> 166    874.642 0.03837683
#> 167    878.005 0.04322640
#> 168    881.367 0.04541639
#> 169    884.728 0.05030751
#> 170    888.087 0.05369016
#> 171    891.445 0.05744265
#> 172    894.802 0.05930020
#> 173    898.157 0.06166057
#> 174    901.511 0.05849402
#> 175    904.864 0.05898022
#> 176    908.216 0.06001080
#> 177    911.567 0.05655336
#> 178    914.914 0.05268035
#> 179    918.262 0.05584275
#> 180    921.609 0.04851230
#> 181    924.954 0.04914395
#> 182    928.296 0.05139628
#> 183    931.639 0.05207779
#> 184    934.981 0.05220246
#> 185    938.320 0.05579289
#> 186    941.659 0.05737616
#> 187    944.995 0.05632896
#> 188    948.332 0.05955369
#> 189    951.665 0.06105801
#> 190    954.999 0.05804106
#> 191    958.330 0.05782081
#> 192    961.660 0.05959109
#> 193    964.991 0.05919631
#> 194    968.318 0.05676529
#> 195    971.644 0.05268035
#> 196    974.971 0.05186586
#> 197    978.295 0.05147523
#> 198    981.617 0.05282580
#> 199    984.940 0.05226895
#> 200    988.258 0.05214428
#> 201    991.578 0.05211104
#> 202    994.896 0.05885971
#> 203    998.212 0.06204289
#> 204   1001.530 0.06178524
#> 205   1004.840 0.05993600
#> 206   1008.150 0.05960771
#> 207   1011.460 0.06053856
#> 208   1014.770 0.06795213
#> 209   1018.080 0.06987616
#> 210   1021.390 0.07081117
#> 211   1024.700 0.07499169
#> 212   1028.000 0.08302028
#> 213   1031.300 0.08475731
#> 214   1034.610 0.08104638
#> 215   1037.910 0.07562334
#> 216   1041.210 0.07298039
#> 217   1044.510 0.07239029
#> 218   1047.800 0.07729388
#> 219   1051.100 0.09511719
#> 220   1054.390 0.11514711
#> 221   1057.680 0.13248836
#> 222   1060.980 0.14271526
#> 223   1064.270 0.14011386
#> 224   1067.560 0.13004488
#> 225   1070.840 0.11597822
#> 226   1074.130 0.10192404
#> 227   1077.420 0.08221410
#> 228   1080.700 0.07554023
#> 229   1083.980 0.07738115
#> 230   1087.270 0.07947557
#> 231   1090.550 0.07842420
#> 232   1093.830 0.07697390
#> 233   1097.100 0.07145113
#> 234   1100.380 0.07145529
#> 235   1103.650 0.07181682
#> 236   1106.930 0.07401097
#> 237   1110.200 0.07221576
#> 238   1113.470 0.07593916
#> 239   1116.740 0.09389129
#> 240   1120.010 0.11479388
#> 241   1123.280 0.12928025
#> 242   1126.550 0.13429604
#> 243   1129.810 0.13275432
#> 244   1133.080 0.12689495
#> 245   1136.340 0.11452377
#> 246   1139.600 0.09630153
#> 247   1142.860 0.08031084
#> 248   1146.120 0.07224069
#> 249   1149.370 0.07692819
#> 250   1152.630 0.07788398
#> 251   1155.890 0.07484624
#> 252   1159.140 0.07844082
#> 253   1162.390 0.08103391
#> 254   1165.640 0.08440409
#> 255   1168.890 0.08781167
#> 256   1172.140 0.08737118
#> 257   1175.390 0.08718833
#> 258   1178.640 0.08592088
#> 259   1181.880 0.07645861
#> 260   1185.120 0.07283910
#> 261   1188.370 0.07214511
#> 262   1191.610 0.07756400
#> 263   1194.850 0.07610123
#> 264   1198.090 0.07534076
#> 265   1201.320 0.07755568
#> 266   1204.560 0.07527842
#> 267   1207.790 0.07140957
#> 268   1211.030 0.07121426
#> 269   1214.260 0.06821809
#> 270   1217.490 0.06729139
#> 271   1220.720 0.06921543
#> 272   1223.950 0.07078624
#> 273   1227.180 0.07283494
#> 274   1230.400 0.07483378
#> 275   1233.630 0.07939245
#> 276   1236.850 0.07615110
#> 277   1240.070 0.07563996
#> 278   1243.300 0.07600981
#> 279   1246.520 0.07683261
#> 280   1249.740 0.07007979
#> 281   1252.950 0.07032912
#> 282   1256.170 0.07133062
#> 283   1259.380 0.07668301
#> 284   1262.600 0.08105884
#> 285   1265.810 0.08105884
#> 286   1269.020 0.08067237
#> 287   1272.230 0.08167387
#> 288   1275.440 0.07994515
#> 289   1278.650 0.07674119
#> 290   1281.850 0.09644697
#> 291   1285.060 0.12923039
#> 292   1288.260 0.15457945
#> 293   1291.460 0.17739362
#> 294   1294.670 0.18769116
#> 295   1297.870 0.18213514
#> 296   1301.070 0.17019199
#> 297   1304.260 0.14379571
#> 298   1307.460 0.10974485
#> 299   1310.660 0.08454953
#> 300   1313.850 0.08772025
#> 301   1317.040 0.08681017
#> 302   1320.230 0.08418384
#> 303   1323.420 0.08676446
#> 304   1326.610 0.08481134
#> 305   1329.800 0.08287899
#> 306   1332.990 0.08065575
#> 307   1336.170 0.08129987
#> 308   1339.360 0.07953790
#> 309   1342.540 0.08173620
#> 310   1345.720 0.07850316
#> 311   1348.900 0.07535738
#> 312   1352.080 0.07135555
#> 313   1355.260 0.07588098
#> 314   1358.440 0.08001995
#> 315   1361.610 0.08156167
#> 316   1364.790 0.08437084
#> 317   1367.960 0.09243684
#> 318   1371.130 0.09294382
#> 319   1374.300 0.09318484
#> 320   1377.470 0.08865525
#> 321   1380.640 0.08960273
#> 322   1383.810 0.08202709
#> 323   1386.970 0.08415060
#> 324   1390.140 0.08920379
#> 325   1393.300 0.08586270
#> 326   1396.470 0.08959441
#> 327   1399.630 0.09069980
#> 328   1402.790 0.09408660
#> 329   1405.950 0.10364860
#> 330   1409.100 0.11029338
#> 331   1412.260 0.11877493
#> 332   1415.410 0.12004654
#> 333   1418.570 0.12802111
#> 334   1421.720 0.13138713
#> 335   1424.870 0.13634890
#> 336   1428.020 0.15250166
#> 337   1431.170 0.17541556
#> 338   1434.320 0.19782663
#> 339   1437.470 0.21326878
#> 340   1440.610 0.21584940
#> 341   1443.760 0.20986536
#> 342   1446.900 0.19749003
#> 343   1450.040 0.18746676
#> 344   1453.180 0.17260638
#> 345   1456.320 0.16800615
#> 346   1459.460 0.17049119
#> 347   1462.590 0.17079455
#> 348   1465.730 0.16203042
#> 349   1468.870 0.15002078
#> 350   1472.000 0.13177360
#> 351   1475.130 0.11661818
#> 352   1478.260 0.10639960
#> 353   1481.400 0.10599651
#> 354   1484.520 0.10559342
#> 355   1487.650 0.10759225
#> 356   1490.780 0.11071310
#> 357   1493.900 0.10729305
#> 358   1497.030 0.10828624
#> 359   1500.150 0.10784159
#> 360   1503.270 0.10148770
#> 361   1506.390 0.09357962
#> 362   1509.510 0.09172623
#> 363   1512.630 0.09143534
#> 364   1515.740 0.09214179
#> 365   1518.860 0.09821310
#> 366   1521.970 0.10933760
#> 367   1525.090 0.11665143
#> 368   1528.200 0.12145529
#> 369   1531.310 0.12546127
#> 370   1534.420 0.12208693
#> 371   1537.530 0.10989860
#> 372   1540.640 0.10212350
#> 373   1543.740 0.09500914
#> 374   1546.850 0.09334275
#> 375   1549.950 0.09429854
#> 376   1553.050 0.09682513
#> 377   1556.160 0.09572806
#> 378   1559.260 0.09771858
#> 379   1562.360 0.10617104
#> 380   1565.450 0.10628324
#> 381   1568.550 0.10602975
#> 382   1571.650 0.10600898
#> 383   1574.740 0.10556433
#> 384   1577.830 0.10506566
#> 385   1580.920 0.10803690
#> 386   1584.020 0.10548953
#> 387   1587.110 0.09847490
#> 388   1590.190 0.09727809
#> 389   1593.280 0.09323471
#> 390   1596.370 0.08821892
#> 391   1599.450 0.08759973
#> 392   1602.540 0.08809425
#> 393   1605.620 0.08919963
#> 394   1608.700 0.09392453
#> 395   1611.780 0.10152094
#> 396   1614.860 0.10302527
#> 397   1617.940 0.10233544
#> 398   1621.020 0.10415974
#> 399   1624.090 0.10138381
#> 400   1627.170 0.09937251
#> 401   1630.240 0.09892370
#> 402   1633.310 0.10448388
#> 403   1636.380 0.10915060
#> 404   1639.450 0.11147773
#> 405   1642.520 0.11159408
#> 406   1645.590 0.11798953
#> 407   1648.650 0.11700881
#> 408   1651.720 0.11755319
#> 409   1654.780 0.11318152
#> 410   1657.850 0.11050947
#> 411   1660.910 0.11199302
#> 412   1663.970 0.11385057
#> 413   1667.030 0.11381316
#> 414   1670.090 0.11208444
#> 415   1673.150 0.11371343
#> 416   1676.200 0.11865027
#> 417   1679.260 0.11934840
#> 418   1682.310 0.11992188
#> 419   1685.360 0.11227975
#> 420   1688.410 0.11185173
#> 421   1691.460 0.11155668
#> 422   1694.510 0.11149435
#> 423   1697.560 0.11054272
#> 424   1700.610 0.11158993
#> 425   1703.650 0.11529671
#> 426   1706.700 0.12207031
#> 427   1709.740 0.13022357
#> 428   1712.780 0.12832862
#> 429   1715.830 0.12468418
#> 430   1718.860 0.12355386
#> 431   1721.900 0.12134724
#> 432   1724.940 0.11715426
#> 433   1727.980 0.11236702
#> 434   1731.010 0.10989029
#> 435   1734.050 0.10783328
#> 436   1737.080 0.11016456
#> 437   1740.110 0.10955369
#> 438   1743.140 0.10863115
#> 439   1746.170 0.10984458
#> 440   1749.200 0.11511386
#> 441   1752.230 0.11685505
#> 442   1755.250 0.11670961
#> 443   1758.280 0.12104804
#> 444   1761.300 0.12030419
#> 445   1764.330 0.11908245
#> 446   1767.350 0.11844249
#> 447   1770.370 0.11357214
#> 448   1773.390 0.11499335
#> 449   1776.410 0.12157164
#> 450   1779.420 0.13193567
#> 451   1782.440 0.13466589
#> 452   1785.460 0.13518118
#> 453   1788.470 0.13825632
#> 454   1791.480 0.13723820
#> 455   1794.490 0.13615359
#> 456   1797.500 0.13126247
#> 457   1800.510 0.12354555
#> 458   1803.520 0.12209525
#> 459   1806.530 0.12685755
#> 460   1809.530 0.12858211
#> 461   1812.540 0.12912650
#> 462   1815.540 0.12447224
#> 463   1818.540 0.12866938
#> 464   1821.540 0.13120844
#> 465   1824.550 0.13191489
#> 466   1827.550 0.13227643
#> 467   1830.540 0.13493185
#> 468   1833.540 0.13700133
#> 469   1836.540 0.14114860
#> 470   1839.530 0.14260306
#> 471   1842.520 0.14403674
#> 472   1845.510 0.14328042
#> 473   1848.510 0.14449385
#> 474   1851.500 0.14232048
#> 475   1854.490 0.13972324
#> 476   1857.480 0.13912068
#> 477   1860.460 0.13934508
#> 478   1863.450 0.13592088
#> 479   1866.430 0.13364777
#> 480   1869.410 0.13280834
#> 481   1872.400 0.13246343
#> 482   1875.380 0.13277510
#> 483   1878.360 0.13404255
#> 484   1881.340 0.13587101
#> 485   1884.320 0.13665642
#> 486   1887.290 0.13665226
#> 487   1890.270 0.13649435
#> 488   1893.240 0.13671044
#> 489   1896.220 0.13321975
#> 490   1899.190 0.13252161
#> 491   1902.160 0.12968750
#> 492   1905.130 0.12681184
#> 493   1908.100 0.12623836
#> 494   1911.070 0.12881898
#> 495   1914.030 0.13094664
#> 496   1917.000 0.13089262
#> 497   1919.960 0.13314910
#> 498   1922.930 0.13595828
#> 499   1925.890 0.13384724
#> 500   1928.850 0.13781167
#> 501   1931.810 0.13305768
#> 502   1934.770 0.12876496
#> 503   1937.430 0.12955868
#> 504   1940.400 0.13596243
#> 505   1943.360 0.14186752
#> 506   1946.330 0.14185090
#> 507   1949.300 0.14099069
#> 508   1952.270 0.14256981
#> 509   1955.230 0.14410322
#> 510   1958.190 0.14585688
#> 511   1961.160 0.13983959
#> 512   1964.120 0.14038398
#> 513   1967.080 0.14079122
#> 514   1970.040 0.14047955
#> 515   1972.990 0.13934924
#> 516   1975.950 0.13499003
#> 517   1978.910 0.13642370
#> 518   1981.860 0.14146027
#> 519   1984.820 0.14222074
#> 520   1987.770 0.14126912
#> 521   1990.720 0.14278175
#> 522   1993.670 0.15013713
#> 523   1996.620 0.14707447
#> 524   1999.570 0.14836686
#> 525   2002.520 0.14630568
#> 526   2005.460 0.14811752
#> 527   2008.410 0.14974651
#> 528   2011.350 0.14960106
#> 529   2014.290 0.14743185
#> 530   2017.240 0.14286486
#> 531   2020.180 0.14247839
#> 532   2023.120 0.14468085
#> 533   2026.050 0.14210439
#> 534   2028.990 0.14564495
#> 535   2031.930 0.14141872
#> 536   2034.870 0.14677527
#> 537   2037.800 0.14940991
#> 538   2040.730 0.15131316
#> 539   2043.660 0.15285489
#> 540   2046.600 0.15126330
#> 541   2049.520 0.14659658
#> 542   2052.450 0.14636386
#> 543   2055.380 0.14852061
#> 544   2058.310 0.15378989
#> 545   2061.230 0.15292553
#> 546   2064.160 0.15664894
#> 547   2067.080 0.16161486
#> 548   2070.000 0.16169797
#> 549   2072.930 0.16315658
#> 550   2075.850 0.16081283
#> 551   2078.770 0.15804106
#> 552   2081.680 0.16276596
#> 553   2084.600 0.17031666
#> 554   2087.520 0.17408577
#> 555   2090.430 0.17062417
#> 556   2093.340 0.16948138
#> 557   2096.260 0.17202045
#> 558   2099.170 0.17339594
#> 559   2102.080 0.17472989
#> 560   2104.990 0.17524934
#> 561   2107.900 0.17760140
#> 562   2110.810 0.17900598
#> 563   2113.710 0.17760555
#> 564   2116.620 0.17473404
#> 565   2119.520 0.16734126
#> 566   2122.430 0.16813497
#> 567   2125.330 0.16564993
#> 568   2128.230 0.16821393
#> 569   2131.130 0.17059508
#> 570   2134.030 0.17268949
#> 571   2136.930 0.17145529
#> 572   2139.820 0.17524934
#> 573   2142.720 0.17585605
#> 574   2145.610 0.18205618
#> 575   2148.510 0.18235539
#> 576   2151.400 0.18336519
#> 577   2154.290 0.18084691
#> 578   2157.180 0.18247174
#> 579   2160.070 0.18056017
#> 580   2162.960 0.17735622
#> 581   2165.850 0.17100233
#> 582   2168.730 0.17273936
#> 583   2171.620 0.17212434
#> 584   2174.500 0.17282247
#> 585   2177.390 0.16845080
#> 586   2180.270 0.16613198
#> 587   2183.150 0.16887882
#> 588   2186.030 0.17491273
#> 589   2188.910 0.17843667
#> 590   2191.790 0.17787982
#> 591   2194.670 0.17372839
#> 592   2197.540 0.17179604
#> 593   2200.420 0.16832613
#> 594   2203.290 0.16666390
#> 595   2206.160 0.16515957
#> 596   2209.030 0.16847573
#> 597   2211.900 0.17063664
#> 598   2214.770 0.17295130
#> 599   2217.640 0.17170878
#> 600   2220.510 0.16789395
#> 601   2223.380 0.16673454
#> 602   2226.240 0.16822640
#> 603   2229.110 0.16651430
#> 604   2231.970 0.17615525
#> 605   2234.830 0.17737284
#> 606   2237.690 0.17440991
#> 607   2240.550 0.17057846
#> 608   2243.410 0.15973654
#> 609   2246.270 0.15437999
#> 610   2249.130 0.15147108
#> 611   2251.980 0.15561004
#> 612   2254.840 0.15856051
#> 613   2257.690 0.16203042
#> 614   2260.540 0.17257314
#> 615   2263.390 0.17079039
#> 616   2266.250 0.16785655
#> 617   2269.100 0.16785239
#> 618   2271.940 0.16006067
#> 619   2274.790 0.15533993
#> 620   2277.640 0.15435921
#> 621   2280.490 0.15826546
#> 622   2283.330 0.16051779
#> 623   2286.170 0.16526346
#> 624   2289.020 0.16673870
#> 625   2291.860 0.16597407
#> 626   2294.700 0.16429521
#> 627   2297.540 0.16241689
#> 628   2300.380 0.15599235
#> 629   2303.210 0.15184508
#> 630   2306.050 0.15034076
#> 631   2308.890 0.14913564
#> 632   2311.720 0.14576546
#> 633   2314.550 0.14698305
#> 634   2317.390 0.14949302
#> 635   2320.220 0.15117188
#> 636   2323.050 0.14930602
#> 637   2325.880 0.15073138
#> 638   2328.700 0.14732380
#> 639   2331.530 0.14734458
#> 640   2334.360 0.14707447
#> 641   2337.180 0.14566157
#> 642   2340.010 0.14129820
#> 643   2342.830 0.14181350
#> 644   2345.650 0.14805519
#> 645   2348.470 0.14991689
#> 646   2351.290 0.15097241
#> 647   2354.110 0.14857463
#> 648   2356.930 0.14380402
#> 649   2359.750 0.13756649
#> 650   2362.560 0.13609541
#> 651   2365.380 0.13213514
#> 652   2368.190 0.12811253
#> 653   2371.000 0.12791307
#> 654   2373.820 0.13457031
#> 655   2376.630 0.13784076
#> 656   2379.440 0.14368351
#> 657   2382.250 0.14185505
#> 658   2385.050 0.14305186
#> 659   2387.860 0.14037151
#> 660   2390.670 0.13924119
#> 661   2393.470 0.13447889
#> 662   2396.280 0.13107131
#> 663   2399.080 0.12792138
#> 664   2401.880 0.12455120
#> 665   2404.680 0.12397357
#> 666   2407.480 0.12420628
#> 667   2410.280 0.11973072
#> 668   2413.080 0.11999252
#> 669   2415.870 0.12285572
#> 670   2418.670 0.12326297
#> 671   2421.460 0.12235705
#> 672   2424.260 0.12616772
#> 673   2427.050 0.12716922
#> 674   2429.840 0.13020695
#> 675   2432.630 0.13422124
#> 676   2435.420 0.13654837
#> 677   2438.210 0.13471160
#> 678   2441.000 0.13591672
#> 679   2443.790 0.13645695
#> 680   2446.570 0.13539727
#> 681   2449.360 0.13143285
#> 682   2452.140 0.13011968
#> 683   2454.920 0.12472989
#> 684   2457.700 0.12341257
#> 685   2460.480 0.12460938
#> 686   2463.260 0.12379488
#> 687   2466.040 0.12107297
#> 688   2468.820 0.12244847
#> 689   2471.600 0.12601396
#> 690   2474.370 0.13030668
#> 691   2477.150 0.13092171
#> 692   2479.920 0.13090093
#> 693   2482.690 0.13230967
#> 694   2485.470 0.13343584
#> 695   2488.240 0.13319066
#> 696   2491.010 0.13017370
#> 697   2493.770 0.12766373
#> 698   2496.540 0.12585605
#> 699   2499.310 0.12615941
#> 700   2502.070 0.12533245
#> 701   2504.840 0.12195396
#> 702   2507.600 0.12124751
#> 703   2510.370 0.12063664
#> 704   2513.130 0.12152178
#> 705   2515.890 0.12045795
#> 706   2518.650 0.12171293
#> 707   2521.410 0.12166307
#> 708   2524.160 0.11887882
#> 709   2526.920 0.12285572
#> 710   2529.680 0.12446809
#> 711   2532.430 0.12554438
#> 712   2535.190 0.13036902
#> 713   2537.940 0.13065160
#> 714   2540.690 0.13183178
#> 715   2543.440 0.13113780
#> 716   2546.190 0.12899767
#> 717   2548.940 0.12668717
#> 718   2551.690 0.12186253
#> 719   2554.440 0.11916140
#> 720   2557.180 0.11628989
#> 721   2559.930 0.11709608
#> 722   2562.670 0.11877493
#> 723   2565.420 0.11603225
#> 724   2568.160 0.11485622
#> 725   2570.900 0.12175033
#> 726   2573.640 0.12056599
#> 727   2576.380 0.11809757
#> 728   2579.110 0.11470662
#> 729   2581.850 0.11136137
#> 730   2584.590 0.11444481
#> 731   2587.320 0.11683843
#> 732   2590.060 0.11658494
#> 733   2592.790 0.11754072
#> 734   2595.520 0.11553358
#> 735   2598.250 0.11866689
#> 736   2600.980 0.11272440
#> 737   2603.720 0.10977394
#> 738   2606.440 0.11113697
#> 739   2609.170 0.11013132
#> 740   2611.900 0.11548787
#> 741   2614.620 0.11509309
#> 742   2617.350 0.11770279
#> 743   2620.070 0.11797706
#> 744   2622.790 0.11028092
#> 745   2625.510 0.10826130
#> 746   2628.230 0.10637467
#> 747   2630.950 0.10829039
#> 748   2633.670 0.10764212
#> 749   2636.390 0.11144448
#> 750   2639.110 0.11692154
#> 751   2641.820 0.11864195
#> 752   2644.540 0.12821642
#> 753   2647.250 0.12757646
#> 754   2649.960 0.12276845
#> 755   2652.680 0.12042055
#> 756   2655.390 0.12115193
#> 757   2658.100 0.11698388
#> 758   2660.810 0.11527178
#> 759   2663.510 0.11847158
#> 760   2666.220 0.11579122
#> 761   2668.930 0.11821809
#> 762   2671.630 0.11861287
#> 763   2674.340 0.11043052
#> 764   2677.040 0.10539811
#> 765   2679.740 0.10465426
#> 766   2682.440 0.10764212
#> 767   2685.140 0.10487866
#> 768   2687.840 0.11420379
#> 769   2690.540 0.12226563
#> 770   2693.240 0.12716922
#> 771   2695.940 0.12801280
#> 772   2698.630 0.12950881
#> 773   2701.330 0.13019448
#> 774   2704.020 0.13475316
#> 775   2706.710 0.13945728
#> 776   2709.410 0.14721576
#> 777   2712.100 0.15056932
#> 778   2714.790 0.16301114
#> 779   2717.470 0.17021277
#> 780   2720.160 0.17026679
#> 781   2722.850 0.16476895
#> 782   2725.530 0.16251247
#> 783   2728.220 0.15188664
#> 784   2730.900 0.14393700
#> 785   2733.590 0.13872590
#> 786   2736.270 0.13004072
#> 787   2738.950 0.12043717
#> 788   2741.630 0.12014628
#> 789   2744.310 0.12116024
#> 790   2746.990 0.11702959
#> 791   2749.670 0.11508893
#> 792   2752.340 0.11414977
#> 793   2755.020 0.11202211
#> 794   2757.690 0.11375914
#> 795   2760.370 0.11684259
#> 796   2763.040 0.11513880
#> 797   2765.710 0.11653092
#> 798   2768.380 0.12117271
#> 799   2771.050 0.12260223
#> 800   2773.720 0.11621925
#> 801   2776.390 0.11518035
#> 802   2779.060 0.11338514
#> 803   2781.720 0.11204289
#> 804   2784.390 0.11300283
#> 805   2787.050 0.11579122
#> 806   2789.710 0.11947307
#> 807   2792.380 0.12026263
#> 808   2795.040 0.12576463
#> 809   2797.700 0.12763880
#> 810   2800.360 0.12682430
#> 811   2803.020 0.12948388
#> 812   2805.680 0.12940492
#> 813   2808.330 0.13425449
#> 814   2810.990 0.13743351
#> 815   2813.640 0.13810256
#> 816   2816.300 0.14004322
#> 817   2818.950 0.14107380
#> 818   2821.600 0.14790143
#> 819   2824.250 0.15322473
#> 820   2826.900 0.15277178
#> 821   2829.550 0.16933594
#> 822   2832.200 0.20698554
#> 823   2834.850 0.27698637
#> 824   2837.490 0.37558178
#> 825   2840.140 0.48959026
#> 826   2842.780 0.60367354
#> 827   2845.430 0.69381233
#> 828   2848.070 0.74793883
#> 829   2850.710 0.75745928
#> 830   2853.350 0.72103557
#> 831   2855.990 0.66325216
#> 832   2858.630 0.59843750
#> 833   2861.270 0.54076629
#> 834   2863.910 0.49859541
#> 835   2866.540 0.46266622
#> 836   2869.180 0.44710356
#> 837   2871.810 0.53823138
#> 838   2874.450 0.68686420
#> 839   2877.080 0.83417969
#> 840   2879.710 0.94339262
#> 841   2882.340 1.00000000
#> 842   2884.970 0.99308511
#> 843   2887.600 0.92214511
#> 844   2890.230 0.79987533
#> 845   2892.860 0.64955951
#> 846   2895.480 0.52242769
#> 847   2898.110 0.48743351
#> 848   2900.730 0.47839927
#> 849   2903.350 0.46668883
#> 850   2905.970 0.44598986
#> 851   2908.600 0.41957281
#> 852   2911.220 0.39112367
#> 853   2913.840 0.36342670
#> 854   2916.460 0.34256150
#> 855   2919.070 0.32808760
#> 856   2921.690 0.32267703
#> 857   2924.310 0.32042470
#> 858   2926.920 0.31903674
#> 859   2929.540 0.31931100
#> 860   2932.150 0.31421626
#> 861   2934.760 0.30703541
#> 862   2937.370 0.28800283
#> 863   2939.980 0.25884724
#> 864   2942.590 0.22543218
#> 865   2945.200 0.19528757
#> 866   2947.810 0.17183760
#> 867   2950.420 0.15256400
#> 868   2953.020 0.13806516
#> 869   2955.630 0.13342753
#> 870   2958.230 0.13481134
#> 871   2960.830 0.13887134
#> 872   2963.440 0.13253408
#> 873   2966.040 0.12855718
#> 874   2968.640 0.12609707
#> 875   2971.240 0.12015459
#> 876   2973.830 0.11583693
#> 877   2976.430 0.11210106
#> 878   2979.030 0.11411237
#> 879   2981.630 0.11563331
#> 880   2984.220 0.11854638
#> 881   2986.810 0.11989694
#> 882   2989.410 0.11805602
#> 883   2992.000 0.11343085
#> 884   2994.590 0.11400017
#> 885   2997.180 0.11224651
#> 886   2999.770 0.11196809
#> 887   3002.360 0.10623753
#> 888   3004.950 0.10605053
#> 889   3007.530 0.10667803
#> 890   3010.120 0.10703956
#> 891   3012.700 0.10340758
#> 892   3015.290 0.10013298
#> 893   3017.870 0.09984624
#> 894   3020.450 0.09968418
#> 895   3023.040 0.09846243
#> 896   3025.620 0.09928108
#> 897   3028.200 0.09591506
#> 898   3030.770 0.09584441
#> 899   3033.350 0.09328873
#> 900   3035.930 0.09360871
#> 901   3038.500 0.09211686
#> 902   3041.080 0.09348404
#> 903   3043.650 0.09525017
#> 904   3046.230 0.09399518
#> 905   3048.800 0.09531666
#> 906   3051.370 0.09982962
#> 907   3053.940 0.09977144
#> 908   3056.510 0.09980469
#> 909   3059.080 0.09676695
#> 910   3061.650 0.09872839
#> 911   3064.210 0.09805934
#> 912   3066.780 0.09743185
#> 913   3069.350 0.09696642
#> 914   3071.910 0.09862866
#> 915   3074.470 0.09960106
#> 916   3077.040 0.09905668
#> 917   3079.600 0.09846243
#> 918   3082.160 0.09985455
#> 919   3084.720 0.09564495
#> 920   3087.280 0.09433178
#> 921   3089.830 0.09275682
#> 922   3092.390 0.09357962
#> 923   3094.950 0.09129405
#> 924   3097.500 0.08988115
#> 925   3100.060 0.09044631
#> 926   3102.610 0.08897939
#> 927   3105.160 0.08479471
#> 928   3107.720 0.08461602
#> 929   3110.270 0.08336104
#> 930   3112.820 0.08461187
#> 931   3115.370 0.08422124
#> 932   3117.910 0.08348570
#> 933   3120.460 0.08108793
#> 934   3123.010 0.07974152
#> 935   3125.560 0.08048122
#> 936   3128.100 0.08070562
#> 937   3130.640 0.07982879
#> 938   3133.190 0.08218085
#> 939   3135.730 0.08375582
#> 940   3138.270 0.08068484
#> 941   3140.810 0.07770113
#> 942   3143.350 0.07626330
#> 943   3145.890 0.07442653
#> 944   3148.430 0.07164644
#> 945   3150.960 0.07281416
#> 946   3153.500 0.07598072
#> 947   3156.030 0.07726479
#> 948   3158.570 0.07811669
#> 949   3161.100 0.07679521
#> 950   3163.640 0.07462184
#> 951   3166.170 0.07481715
#> 952   3168.700 0.07521193
#> 953   3171.230 0.07594747
#> 954   3173.760 0.07491689
#> 955   3175.300 0.07765957
#> 956   3177.840 0.07900598
#> 957   3180.380 0.07677443
#> 958   3182.910 0.07512467
#> 959   3185.450 0.07517869
#> 960   3187.990 0.07476313
#> 961   3190.520 0.07397773
#> 962   3193.060 0.07257314
#> 963   3195.590 0.07030003
#> 964   3198.120 0.06690908