Automated background subtraction for spectral data

This baseline correction routine iteratively finds the baseline of a spectrum using a polynomial fitting.

Usage

subtr_bg(x, ...)

# S3 method for formula
subtr_bg(formula, data = NULL, ...)

# S3 method for data.frame
subtr_bg(x, ...)

# S3 method for default
subtr_bg(x, y, degree = 8, raw = FALSE, make_rel = TRUE, ...)

Arguments

x

a numeric vector containing the spectral wavenumbers; alternatively a data frame containing spectral data as "wavenumber" and "intensity" can be supplied.

formula

an object of class 'formula' of the form intensity ~ wavenumber.

data

a data frame containing the variables in formula.

y

a numeric vector containing the spectral intensities.

degree

the degree of the polynomial. Must be less than the number of unique points when raw is FALSE. Typically a good fit can be found with a 8th order polynomial.

raw

if TRUE, use raw and not orthogonal polynomials.

make_rel

logical; if TRUE spectra are automatically normalized with make_rel().

...

further arguments passed to poly().

Value

subtr_bg() returns a data frame containing two columns named "wavenumber" and "intensity".

Details

This is a translation of Michael Stephen Chen's MATLAB code written for the imodpolyfit routine from Zhao et al. 2007.

References

Chen MS (2020). Michaelstchen/ModPolyFit. MATLAB. Retrieved from https://github.com/michaelstchen/modPolyFit (Original work published July 28, 2015)

Zhao J, Lui H, McLean DI, Zeng H (2007). “Automated Autofluorescence Background Subtraction Algorithm for Biomedical Raman Spectroscopy.” Applied Spectroscopy, 61(11), 1225–1232. doi:10.1366/000370207782597003 .

See also

Author

Win Cowger, Zacharias Steinmetz

Examples

data("raman_hdpe")

subtr_bg(raman_hdpe)
#>     wavenumber   intensity
#> 1      301.040 0.000000000
#> 2      304.632 0.010833354
#> 3      308.221 0.008185951
#> 4      311.810 0.000000000
#> 5      315.398 0.000000000
#> 6      318.983 0.000000000
#> 7      322.566 0.000000000
#> 8      326.150 0.000000000
#> 9      329.732 0.000000000
#> 10     333.311 0.000000000
#> 11     336.889 0.000000000
#> 12     340.467 0.000000000
#> 13     344.042 0.000000000
#> 14     347.618 0.006689701
#> 15     351.190 0.000000000
#> 16     354.760 0.007083756
#> 17     358.332 0.005947830
#> 18     361.900 0.000000000
#> 19     365.466 0.000000000
#> 20     369.031 0.000000000
#> 21     372.597 0.014434510
#> 22     376.160 0.008590664
#> 23     379.721 0.000000000
#> 24     383.280 0.000000000
#> 25     386.839 0.006405977
#> 26     390.396 0.000000000
#> 27     393.951 0.000000000
#> 28     397.505 0.000000000
#> 29     401.058 0.011752738
#> 30     404.610 0.007139497
#> 31     408.160 0.000000000
#> 32     411.708 0.008539468
#> 33     415.256 0.000000000
#> 34     418.801 0.000000000
#> 35     422.346 0.007957459
#> 36     425.887 0.007337918
#> 37     429.429 0.007179291
#> 38     432.970 0.030430612
#> 39     436.507 0.019442180
#> 40     440.044 0.035717009
#> 41     443.581 0.016393826
#> 42     447.116 0.028546737
#> 43     450.647 0.007810273
#> 44     454.180 0.022988930
#> 45     457.708 0.013175104
#> 46     461.238 0.000000000
#> 47     464.764 0.012997979
#> 48     468.291 0.011958591
#> 49     471.815 0.009112736
#> 50     475.339 0.000000000
#> 51     478.859 0.000000000
#> 52     482.379 0.000000000
#> 53     485.899 0.000000000
#> 54     489.416 0.006425891
#> 55     492.931 0.006384515
#> 56     496.447 0.000000000
#> 57     499.960 0.000000000
#> 58     503.471 0.000000000
#> 59     506.981 0.000000000
#> 60     510.489 0.000000000
#> 61     513.999 0.000000000
#> 62     517.503 0.006645212
#> 63     521.009 0.000000000
#> 64     524.510 0.000000000
#> 65     528.014 0.000000000
#> 66     531.513 0.000000000
#> 67     535.014 0.000000000
#> 68     538.510 0.006543985
#> 69     542.008 0.000000000
#> 70     545.502 0.000000000
#> 71     548.997 0.000000000
#> 72     552.488 0.010752349
#> 73     555.979 0.009753444
#> 74     559.469 0.009737342
#> 75     562.955 0.000000000
#> 76     566.442 0.000000000
#> 77     569.928 0.000000000
#> 78     573.411 0.012058656
#> 79     576.894 0.009649910
#> 80     580.376 0.000000000
#> 81     583.854 0.000000000
#> 82     587.331 0.007201601
#> 83     590.811 0.027527479
#> 84     594.286 0.010928313
#> 85     597.759 0.010902715
#> 86     601.231 0.014678866
#> 87     604.705 0.032965241
#> 88     608.174 0.028769327
#> 89     611.642 0.028735160
#> 90     615.109 0.035638343
#> 91     618.574 0.013146016
#> 92     622.038 0.013111524
#> 93     625.503 0.014462866
#> 94     628.966 0.007613028
#> 95     632.426 0.006779146
#> 96     635.884 0.000000000
#> 97     639.342 0.000000000
#> 98     642.798 0.000000000
#> 99     646.252 0.008782931
#> 100    649.706 0.006532741
#> 101    653.158 0.000000000
#> 102    656.608 0.006419347
#> 103    660.058 0.006359363
#> 104    663.506 0.000000000
#> 105    666.953 0.000000000
#> 106    670.398 0.000000000
#> 107    673.842 0.008318654
#> 108    677.285 0.000000000
#> 109    680.727 0.011118859
#> 110    684.165 0.022226062
#> 111    687.602 0.000000000
#> 112    691.040 0.010884593
#> 113    694.476 0.000000000
#> 114    697.911 0.000000000
#> 115    701.345 0.009093142
#> 116    704.775 0.000000000
#> 117    708.204 0.008916488
#> 118    711.634 0.000000000
#> 119    715.063 0.011645747
#> 120    718.490 0.000000000
#> 121    721.912 0.000000000
#> 122    725.336 0.011022242
#> 123    728.760 0.000000000
#> 124    732.180 0.000000000
#> 125    735.599 0.000000000
#> 126    739.018 0.006326305
#> 127    742.434 0.000000000
#> 128    745.849 0.000000000
#> 129    749.265 0.000000000
#> 130    752.675 0.006684051
#> 131    756.088 0.008816234
#> 132    759.498 0.000000000
#> 133    762.906 0.000000000
#> 134    766.313 0.007055924
#> 135    769.719 0.006927863
#> 136    773.124 0.014881717
#> 137    776.527 0.008053922
#> 138    779.929 0.006239599
#> 139    783.330 0.000000000
#> 140    786.729 0.000000000
#> 141    790.127 0.014180884
#> 142    793.524 0.006513817
#> 143    796.919 0.010103664
#> 144    800.312 0.000000000
#> 145    803.705 0.000000000
#> 146    807.097 0.011052016
#> 147    810.485 0.000000000
#> 148    813.874 0.007878417
#> 149    817.262 0.000000000
#> 150    820.647 0.000000000
#> 151    824.032 0.000000000
#> 152    827.416 0.000000000
#> 153    830.797 0.000000000
#> 154    834.179 0.000000000
#> 155    837.557 0.011044933
#> 156    840.936 0.007197467
#> 157    844.312 0.000000000
#> 158    847.688 0.000000000
#> 159    851.061 0.000000000
#> 160    854.433 0.000000000
#> 161    857.806 0.019306672
#> 162    861.175 0.007028987
#> 163    864.543 0.000000000
#> 164    867.910 0.009555058
#> 165    871.278 0.000000000
#> 166    874.642 0.000000000
#> 167    878.005 0.000000000
#> 168    881.367 0.000000000
#> 169    884.728 0.000000000
#> 170    888.087 0.000000000
#> 171    891.445 0.014756635
#> 172    894.802 0.000000000
#> 173    898.157 0.013004688
#> 174    901.511 0.000000000
#> 175    904.864 0.014025950
#> 176    908.216 0.005946084
#> 177    911.567 0.000000000
#> 178    914.914 0.000000000
#> 179    918.262 0.009704207
#> 180    921.609 0.000000000
#> 181    924.954 0.000000000
#> 182    928.296 0.010553020
#> 183    931.639 0.005951655
#> 184    934.981 0.000000000
#> 185    938.320 0.008623309
#> 186    941.659 0.006303201
#> 187    944.995 0.000000000
#> 188    948.332 0.000000000
#> 189    951.665 0.000000000
#> 190    954.999 0.000000000
#> 191    958.330 0.012423577
#> 192    961.660 0.000000000
#> 193    964.991 0.000000000
#> 194    968.318 0.000000000
#> 195    971.644 0.000000000
#> 196    974.971 0.006432495
#> 197    978.295 0.000000000
#> 198    981.617 0.000000000
#> 199    984.940 0.000000000
#> 200    988.258 0.000000000
#> 201    991.578 0.000000000
#> 202    994.896 0.011083808
#> 203    998.212 0.000000000
#> 204   1001.530 0.000000000
#> 205   1004.840 0.006236565
#> 206   1008.150 0.010983710
#> 207   1011.460 0.000000000
#> 208   1014.770 0.000000000
#> 209   1018.080 0.000000000
#> 210   1021.390 0.011023990
#> 211   1024.700 0.028314929
#> 212   1028.000 0.012638116
#> 213   1031.300 0.006231645
#> 214   1034.610 0.011693079
#> 215   1037.910 0.020408169
#> 216   1041.210 0.011896228
#> 217   1044.510 0.000000000
#> 218   1047.800 0.000000000
#> 219   1051.100 0.018282724
#> 220   1054.390 0.032330145
#> 221   1057.680 0.040488233
#> 222   1060.980 0.084397781
#> 223   1064.270 0.104715720
#> 224   1067.560 0.045927292
#> 225   1070.840 0.014412120
#> 226   1074.130 0.009566642
#> 227   1077.420 0.006619912
#> 228   1080.700 0.019420949
#> 229   1083.980 0.010914184
#> 230   1087.270 0.016285530
#> 231   1090.550 0.000000000
#> 232   1093.830 0.012685043
#> 233   1097.100 0.000000000
#> 234   1100.380 0.011408384
#> 235   1103.650 0.005642807
#> 236   1106.930 0.007462945
#> 237   1110.200 0.000000000
#> 238   1113.470 0.012089322
#> 239   1116.740 0.011483116
#> 240   1120.010 0.013126092
#> 241   1123.280 0.034668807
#> 242   1126.550 0.067807436
#> 243   1129.810 0.107886168
#> 244   1133.080 0.032774411
#> 245   1136.340 0.011853904
#> 246   1139.600 0.000000000
#> 247   1142.860 0.017631541
#> 248   1146.120 0.007123301
#> 249   1149.370 0.006956933
#> 250   1152.630 0.006790304
#> 251   1155.890 0.017013005
#> 252   1159.140 0.000000000
#> 253   1162.390 0.000000000
#> 254   1165.640 0.010350864
#> 255   1168.890 0.008797427
#> 256   1172.140 0.040493701
#> 257   1175.390 0.000000000
#> 258   1178.640 0.017329479
#> 259   1181.880 0.006074961
#> 260   1185.120 0.000000000
#> 261   1188.370 0.000000000
#> 262   1191.610 0.006512844
#> 263   1194.850 0.000000000
#> 264   1198.090 0.023358797
#> 265   1201.320 0.009929684
#> 266   1204.560 0.000000000
#> 267   1207.790 0.000000000
#> 268   1211.030 0.000000000
#> 269   1214.260 0.009277126
#> 270   1217.490 0.000000000
#> 271   1220.720 0.000000000
#> 272   1223.950 0.007305089
#> 273   1227.180 0.000000000
#> 274   1230.400 0.000000000
#> 275   1233.630 0.008301394
#> 276   1236.850 0.010914818
#> 277   1240.070 0.006525814
#> 278   1243.300 0.000000000
#> 279   1246.520 0.000000000
#> 280   1249.740 0.000000000
#> 281   1252.950 0.008714847
#> 282   1256.170 0.000000000
#> 283   1259.380 0.006975730
#> 284   1262.600 0.000000000
#> 285   1265.810 0.010839834
#> 286   1269.020 0.000000000
#> 287   1272.230 0.018498176
#> 288   1275.440 0.000000000
#> 289   1278.650 0.000000000
#> 290   1281.850 0.018040388
#> 291   1285.060 0.008440613
#> 292   1288.260 0.034646589
#> 293   1291.460 0.071954942
#> 294   1294.670 0.171714582
#> 295   1297.870 0.092444534
#> 296   1301.070 0.039543276
#> 297   1304.260 0.033828139
#> 298   1307.460 0.012872283
#> 299   1310.660 0.000000000
#> 300   1313.850 0.017880245
#> 301   1317.040 0.000000000
#> 302   1320.230 0.000000000
#> 303   1323.420 0.027122788
#> 304   1326.610 0.000000000
#> 305   1329.800 0.000000000
#> 306   1332.990 0.007460990
#> 307   1336.170 0.007296520
#> 308   1339.360 0.000000000
#> 309   1342.540 0.006965269
#> 310   1345.720 0.000000000
#> 311   1348.900 0.012079064
#> 312   1352.080 0.000000000
#> 313   1355.260 0.000000000
#> 314   1358.440 0.000000000
#> 315   1361.610 0.007216394
#> 316   1364.790 0.000000000
#> 317   1367.960 0.031047445
#> 318   1371.130 0.006833952
#> 319   1374.300 0.000000000
#> 320   1377.470 0.018044000
#> 321   1380.640 0.000000000
#> 322   1383.810 0.000000000
#> 323   1386.970 0.007354624
#> 324   1390.140 0.000000000
#> 325   1393.300 0.008244723
#> 326   1396.470 0.000000000
#> 327   1399.630 0.018209186
#> 328   1402.790 0.020807187
#> 329   1405.950 0.000000000
#> 330   1409.100 0.019061996
#> 331   1412.260 0.018881707
#> 332   1415.410 0.040290859
#> 333   1418.570 0.042883029
#> 334   1421.720 0.030208822
#> 335   1424.870 0.034187761
#> 336   1428.020 0.039553695
#> 337   1431.170 0.060184924
#> 338   1434.320 0.057222171
#> 339   1437.470 0.120873123
#> 340   1440.610 0.133174930
#> 341   1443.760 0.116331685
#> 342   1446.900 0.064793266
#> 343   1450.040 0.047948406
#> 344   1453.180 0.046368785
#> 345   1456.320 0.061442227
#> 346   1459.460 0.073739090
#> 347   1462.590 0.083259988
#> 348   1465.730 0.060859225
#> 349   1468.870 0.042621018
#> 350   1472.000 0.026394836
#> 351   1475.130 0.000000000
#> 352   1478.260 0.011490247
#> 353   1481.400 0.021632989
#> 354   1484.520 0.006166122
#> 355   1487.650 0.019841991
#> 356   1490.780 0.012073309
#> 357   1493.900 0.016658865
#> 358   1497.030 0.000000000
#> 359   1500.150 0.011461061
#> 360   1503.270 0.009867454
#> 361   1506.390 0.000000000
#> 362   1509.510 0.009453026
#> 363   1512.630 0.000000000
#> 364   1515.740 0.000000000
#> 365   1518.860 0.000000000
#> 366   1521.970 0.028653443
#> 367   1525.090 0.011177119
#> 368   1528.200 0.028223338
#> 369   1531.310 0.028006992
#> 370   1534.420 0.025013893
#> 371   1537.530 0.000000000
#> 372   1540.640 0.000000000
#> 373   1543.740 0.022967927
#> 374   1546.850 0.000000000
#> 375   1549.950 0.000000000
#> 376   1553.050 0.000000000
#> 377   1556.160 0.010396095
#> 378   1559.260 0.007464914
#> 379   1562.360 0.000000000
#> 380   1565.450 0.000000000
#> 381   1568.550 0.000000000
#> 382   1571.650 0.018163411
#> 383   1574.740 0.006221774
#> 384   1577.830 0.000000000
#> 385   1580.920 0.011376046
#> 386   1584.020 0.000000000
#> 387   1587.110 0.000000000
#> 388   1590.190 0.006542719
#> 389   1593.280 0.009070245
#> 390   1596.370 0.000000000
#> 391   1599.450 0.000000000
#> 392   1602.540 0.000000000
#> 393   1605.620 0.000000000
#> 394   1608.700 0.000000000
#> 395   1611.780 0.000000000
#> 396   1614.860 0.008818508
#> 397   1617.940 0.000000000
#> 398   1621.020 0.000000000
#> 399   1624.090 0.000000000
#> 400   1627.170 0.000000000
#> 401   1630.240 0.000000000
#> 402   1633.310 0.000000000
#> 403   1636.380 0.000000000
#> 404   1639.450 0.006831788
#> 405   1642.520 0.008046883
#> 406   1645.590 0.017269934
#> 407   1648.650 0.000000000
#> 408   1651.720 0.000000000
#> 409   1654.780 0.007070986
#> 410   1657.850 0.023188644
#> 411   1660.910 0.000000000
#> 412   1663.970 0.009108342
#> 413   1667.030 0.000000000
#> 414   1670.090 0.008633352
#> 415   1673.150 0.012549713
#> 416   1676.200 0.000000000
#> 417   1679.260 0.000000000
#> 418   1682.310 0.006576940
#> 419   1685.360 0.008782882
#> 420   1688.410 0.000000000
#> 421   1691.460 0.006869708
#> 422   1694.510 0.000000000
#> 423   1697.560 0.000000000
#> 424   1700.610 0.018151027
#> 425   1703.650 0.000000000
#> 426   1706.700 0.011197359
#> 427   1709.740 0.000000000
#> 428   1712.780 0.010695889
#> 429   1715.830 0.011831732
#> 430   1718.860 0.020723040
#> 431   1721.900 0.000000000
#> 432   1724.940 0.000000000
#> 433   1727.980 0.000000000
#> 434   1731.010 0.000000000
#> 435   1734.050 0.000000000
#> 436   1737.080 0.000000000
#> 437   1740.110 0.000000000
#> 438   1743.140 0.000000000
#> 439   1746.170 0.000000000
#> 440   1749.200 0.000000000
#> 441   1752.230 0.000000000
#> 442   1755.250 0.006789392
#> 443   1758.280 0.000000000
#> 444   1761.300 0.008056609
#> 445   1764.330 0.000000000
#> 446   1767.350 0.007493120
#> 447   1770.370 0.000000000
#> 448   1773.390 0.000000000
#> 449   1776.410 0.000000000
#> 450   1779.420 0.000000000
#> 451   1782.440 0.000000000
#> 452   1785.460 0.028188959
#> 453   1788.470 0.029842262
#> 454   1791.480 0.008259521
#> 455   1794.490 0.000000000
#> 456   1797.500 0.000000000
#> 457   1800.510 0.016245879
#> 458   1803.520 0.005866078
#> 459   1806.530 0.000000000
#> 460   1809.530 0.000000000
#> 461   1812.540 0.000000000
#> 462   1815.540 0.007189629
#> 463   1818.540 0.008895202
#> 464   1821.540 0.000000000
#> 465   1824.550 0.006433477
#> 466   1827.550 0.000000000
#> 467   1830.540 0.011037250
#> 468   1833.540 0.000000000
#> 469   1836.540 0.007351361
#> 470   1839.530 0.011708946
#> 471   1842.520 0.012860108
#> 472   1845.510 0.006481614
#> 473   1848.510 0.012051338
#> 474   1851.500 0.009040957
#> 475   1854.490 0.007313287
#> 476   1857.480 0.008574194
#> 477   1860.460 0.000000000
#> 478   1863.450 0.000000000
#> 479   1866.430 0.006155888
#> 480   1869.410 0.005915609
#> 481   1872.400 0.000000000
#> 482   1875.380 0.000000000
#> 483   1878.360 0.000000000
#> 484   1881.340 0.007385118
#> 485   1884.320 0.000000000
#> 486   1887.290 0.000000000
#> 487   1890.270 0.007351364
#> 488   1893.240 0.000000000
#> 489   1896.220 0.000000000
#> 490   1899.190 0.000000000
#> 491   1902.160 0.000000000
#> 492   1905.130 0.000000000
#> 493   1908.100 0.000000000
#> 494   1911.070 0.000000000
#> 495   1914.030 0.000000000
#> 496   1917.000 0.000000000
#> 497   1919.960 0.000000000
#> 498   1922.930 0.009173324
#> 499   1925.890 0.000000000
#> 500   1928.850 0.000000000
#> 501   1931.810 0.000000000
#> 502   1934.770 0.000000000
#> 503   1937.430 0.000000000
#> 504   1940.400 0.006120902
#> 505   1943.360 0.000000000
#> 506   1946.330 0.005877351
#> 507   1949.300 0.016258492
#> 508   1952.270 0.011546265
#> 509   1955.230 0.000000000
#> 510   1958.190 0.000000000
#> 511   1961.160 0.000000000
#> 512   1964.120 0.012203725
#> 513   1967.080 0.000000000
#> 514   1970.040 0.000000000
#> 515   1972.990 0.000000000
#> 516   1975.950 0.007181561
#> 517   1978.910 0.000000000
#> 518   1981.860 0.000000000
#> 519   1984.820 0.000000000
#> 520   1987.770 0.007045163
#> 521   1990.720 0.015219203
#> 522   1993.670 0.000000000
#> 523   1996.620 0.000000000
#> 524   1999.570 0.000000000
#> 525   2002.520 0.012833825
#> 526   2005.460 0.012684564
#> 527   2008.410 0.000000000
#> 528   2011.350 0.012777779
#> 529   2014.290 0.000000000
#> 530   2017.240 0.009333720
#> 531   2020.180 0.000000000
#> 532   2023.120 0.007500781
#> 533   2026.050 0.000000000
#> 534   2028.990 0.000000000
#> 535   2031.930 0.007108341
#> 536   2034.870 0.000000000
#> 537   2037.800 0.000000000
#> 538   2040.730 0.000000000
#> 539   2043.660 0.000000000
#> 540   2046.600 0.022349680
#> 541   2049.520 0.000000000
#> 542   2052.450 0.000000000
#> 543   2055.380 0.000000000
#> 544   2058.310 0.000000000
#> 545   2061.230 0.000000000
#> 546   2064.160 0.025888567
#> 547   2067.080 0.020241110
#> 548   2070.000 0.011403793
#> 549   2072.930 0.000000000
#> 550   2075.850 0.008273089
#> 551   2078.770 0.019888508
#> 552   2081.680 0.008278983
#> 553   2084.600 0.012789913
#> 554   2087.520 0.012284284
#> 555   2090.430 0.026665849
#> 556   2093.340 0.018123018
#> 557   2096.260 0.016667900
#> 558   2099.170 0.006667404
#> 559   2102.080 0.012378773
#> 560   2104.990 0.025030218
#> 561   2107.900 0.023600768
#> 562   2110.810 0.023561803
#> 563   2113.710 0.017713212
#> 564   2116.620 0.017666135
#> 565   2119.520 0.011617130
#> 566   2122.430 0.012029505
#> 567   2125.330 0.008757358
#> 568   2128.230 0.018895106
#> 569   2131.130 0.007134078
#> 570   2134.030 0.030285999
#> 571   2136.930 0.007075488
#> 572   2139.820 0.027483504
#> 573   2142.720 0.014598967
#> 574   2145.610 0.013193469
#> 575   2148.510 0.018730016
#> 576   2151.400 0.038564750
#> 577   2154.290 0.021485067
#> 578   2157.180 0.031631033
#> 579   2160.070 0.011274006
#> 580   2162.960 0.009883663
#> 581   2165.850 0.026107998
#> 582   2168.730 0.018709536
#> 583   2171.620 0.013167114
#> 584   2174.500 0.014566764
#> 585   2177.390 0.009915214
#> 586   2180.270 0.023432782
#> 587   2183.150 0.009948352
#> 588   2186.030 0.008447319
#> 589   2188.910 0.014670101
#> 590   2191.790 0.031887675
#> 591   2194.670 0.024987067
#> 592   2197.540 0.025028080
#> 593   2200.420 0.006106843
#> 594   2203.290 0.000000000
#> 595   2206.160 0.008809722
#> 596   2209.030 0.013549435
#> 597   2211.900 0.023884148
#> 598   2214.770 0.028104051
#> 599   2217.640 0.012319370
#> 600   2220.510 0.010989867
#> 601   2223.380 0.000000000
#> 602   2226.240 0.009077703
#> 603   2229.110 0.016733034
#> 604   2231.970 0.021555510
#> 605   2234.830 0.027180393
#> 606   2237.690 0.019725389
#> 607   2240.550 0.000000000
#> 608   2243.410 0.034354649
#> 609   2246.270 0.000000000
#> 610   2249.130 0.000000000
#> 611   2251.980 0.011242474
#> 612   2254.840 0.000000000
#> 613   2257.690 0.025078070
#> 614   2260.540 0.016259803
#> 615   2263.390 0.022493277
#> 616   2266.250 0.008936916
#> 617   2269.100 0.020728878
#> 618   2271.940 0.006715499
#> 619   2274.790 0.000000000
#> 620   2277.640 0.010747281
#> 621   2280.490 0.011456796
#> 622   2283.330 0.000000000
#> 623   2286.170 0.008314838
#> 624   2289.020 0.027621744
#> 625   2291.860 0.018877908
#> 626   2294.700 0.010677385
#> 627   2297.540 0.008789971
#> 628   2300.380 0.008922960
#> 629   2303.210 0.007670057
#> 630   2306.050 0.000000000
#> 631   2308.890 0.000000000
#> 632   2311.720 0.005648136
#> 633   2314.550 0.009614445
#> 634   2317.390 0.000000000
#> 635   2320.220 0.000000000
#> 636   2323.050 0.000000000
#> 637   2325.880 0.020544195
#> 638   2328.700 0.006675567
#> 639   2331.530 0.000000000
#> 640   2334.360 0.000000000
#> 641   2337.180 0.008037037
#> 642   2340.010 0.000000000
#> 643   2342.830 0.012520348
#> 644   2345.650 0.011295069
#> 645   2348.470 0.000000000
#> 646   2351.290 0.000000000
#> 647   2354.110 0.010406219
#> 648   2356.930 0.022571266
#> 649   2359.750 0.000000000
#> 650   2362.560 0.000000000
#> 651   2365.380 0.000000000
#> 652   2368.190 0.000000000
#> 653   2371.000 0.000000000
#> 654   2373.820 0.008854788
#> 655   2376.630 0.000000000
#> 656   2379.440 0.000000000
#> 657   2382.250 0.000000000
#> 658   2385.050 0.000000000
#> 659   2387.860 0.006292644
#> 660   2390.670 0.000000000
#> 661   2393.470 0.000000000
#> 662   2396.280 0.000000000
#> 663   2399.080 0.000000000
#> 664   2401.880 0.000000000
#> 665   2404.680 0.000000000
#> 666   2407.480 0.000000000
#> 667   2410.280 0.000000000
#> 668   2413.080 0.000000000
#> 669   2415.870 0.000000000
#> 670   2418.670 0.000000000
#> 671   2421.460 0.000000000
#> 672   2424.260 0.000000000
#> 673   2427.050 0.000000000
#> 674   2429.840 0.000000000
#> 675   2432.630 0.000000000
#> 676   2435.420 0.007791996
#> 677   2438.210 0.000000000
#> 678   2441.000 0.010748522
#> 679   2443.790 0.000000000
#> 680   2446.570 0.000000000
#> 681   2449.360 0.000000000
#> 682   2452.140 0.000000000
#> 683   2454.920 0.000000000
#> 684   2457.700 0.000000000
#> 685   2460.480 0.000000000
#> 686   2463.260 0.000000000
#> 687   2466.040 0.000000000
#> 688   2468.820 0.000000000
#> 689   2471.600 0.000000000
#> 690   2474.370 0.000000000
#> 691   2477.150 0.000000000
#> 692   2479.920 0.009479601
#> 693   2482.690 0.009686595
#> 694   2485.470 0.000000000
#> 695   2488.240 0.000000000
#> 696   2491.010 0.000000000
#> 697   2493.770 0.010512913
#> 698   2496.540 0.000000000
#> 699   2499.310 0.000000000
#> 700   2502.070 0.005951552
#> 701   2504.840 0.000000000
#> 702   2507.600 0.000000000
#> 703   2510.370 0.000000000
#> 704   2513.130 0.000000000
#> 705   2515.890 0.000000000
#> 706   2518.650 0.000000000
#> 707   2521.410 0.000000000
#> 708   2524.160 0.008581265
#> 709   2526.920 0.000000000
#> 710   2529.680 0.000000000
#> 711   2532.430 0.000000000
#> 712   2535.190 0.000000000
#> 713   2537.940 0.024180002
#> 714   2540.690 0.000000000
#> 715   2543.440 0.007633675
#> 716   2546.190 0.011527492
#> 717   2548.940 0.000000000
#> 718   2551.690 0.000000000
#> 719   2554.440 0.007025070
#> 720   2557.180 0.000000000
#> 721   2559.930 0.000000000
#> 722   2562.670 0.000000000
#> 723   2565.420 0.000000000
#> 724   2568.160 0.008841102
#> 725   2570.900 0.000000000
#> 726   2573.640 0.000000000
#> 727   2576.380 0.000000000
#> 728   2579.110 0.008153363
#> 729   2581.850 0.008865101
#> 730   2584.590 0.000000000
#> 731   2587.320 0.000000000
#> 732   2590.060 0.007995069
#> 733   2592.790 0.006136356
#> 734   2595.520 0.008329804
#> 735   2598.250 0.000000000
#> 736   2600.980 0.000000000
#> 737   2603.720 0.000000000
#> 738   2606.440 0.000000000
#> 739   2609.170 0.009130526
#> 740   2611.900 0.000000000
#> 741   2614.620 0.006984030
#> 742   2617.350 0.000000000
#> 743   2620.070 0.000000000
#> 744   2622.790 0.008486161
#> 745   2625.510 0.000000000
#> 746   2628.230 0.006408482
#> 747   2630.950 0.000000000
#> 748   2633.670 0.000000000
#> 749   2636.390 0.000000000
#> 750   2639.110 0.007013931
#> 751   2641.820 0.007158355
#> 752   2644.540 0.000000000
#> 753   2647.250 0.026869462
#> 754   2649.960 0.000000000
#> 755   2652.680 0.010618293
#> 756   2655.390 0.012114228
#> 757   2658.100 0.000000000
#> 758   2660.810 0.000000000
#> 759   2663.510 0.009646163
#> 760   2666.220 0.008355462
#> 761   2668.930 0.000000000
#> 762   2671.630 0.019669823
#> 763   2674.340 0.000000000
#> 764   2677.040 0.000000000
#> 765   2679.740 0.000000000
#> 766   2682.440 0.000000000
#> 767   2685.140 0.000000000
#> 768   2687.840 0.006972010
#> 769   2690.540 0.009821934
#> 770   2693.240 0.009167182
#> 771   2695.940 0.006739728
#> 772   2698.630 0.035905902
#> 773   2701.330 0.010741144
#> 774   2704.020 0.006236133
#> 775   2706.710 0.006980539
#> 776   2709.410 0.044208266
#> 777   2712.100 0.032110667
#> 778   2714.790 0.041523703
#> 779   2717.470 0.037401142
#> 780   2720.160 0.047154068
#> 781   2722.850 0.034698719
#> 782   2725.530 0.049996900
#> 783   2728.220 0.029515029
#> 784   2730.900 0.023990519
#> 785   2733.590 0.021238618
#> 786   2736.270 0.008460847
#> 787   2738.950 0.000000000
#> 788   2741.630 0.019903068
#> 789   2744.310 0.011279865
#> 790   2746.990 0.000000000
#> 791   2749.670 0.007903299
#> 792   2752.340 0.018541735
#> 793   2755.020 0.006387778
#> 794   2757.690 0.000000000
#> 795   2760.370 0.007902459
#> 796   2763.040 0.009282945
#> 797   2765.710 0.007813141
#> 798   2768.380 0.007815110
#> 799   2771.050 0.007856415
#> 800   2773.720 0.014058062
#> 801   2776.390 0.006415849
#> 802   2779.060 0.000000000
#> 803   2781.720 0.000000000
#> 804   2784.390 0.000000000
#> 805   2787.050 0.013985325
#> 806   2789.710 0.000000000
#> 807   2792.380 0.015324303
#> 808   2795.040 0.019459362
#> 809   2797.700 0.013876967
#> 810   2800.360 0.006992661
#> 811   2803.020 0.008858739
#> 812   2805.680 0.031873430
#> 813   2808.330 0.008762652
#> 814   2810.990 0.016459761
#> 815   2813.640 0.027544194
#> 816   2816.300 0.024709185
#> 817   2818.950 0.034363027
#> 818   2821.600 0.019040001
#> 819   2824.250 0.030072343
#> 820   2826.900 0.050823753
#> 821   2829.550 0.052144001
#> 822   2832.200 0.074280273
#> 823   2834.850 0.103354350
#> 824   2837.490 0.146305751
#> 825   2840.140 0.289180871
#> 826   2842.780 0.393196992
#> 827   2845.430 0.495825013
#> 828   2848.070 0.547103557
#> 829   2850.710 0.513724410
#> 830   2853.350 0.466466953
#> 831   2855.990 0.406719017
#> 832   2858.630 0.362237221
#> 833   2861.270 0.330245905
#> 834   2863.910 0.309357237
#> 835   2866.540 0.291244501
#> 836   2869.180 0.295336790
#> 837   2871.810 0.291102364
#> 838   2874.450 0.332666091
#> 839   2877.080 0.432518989
#> 840   2879.710 0.666991494
#> 841   2882.340 1.000000000
#> 842   2884.970 0.788978755
#> 843   2887.600 0.552976552
#> 844   2890.230 0.412734688
#> 845   2892.860 0.355762683
#> 846   2895.480 0.308505766
#> 847   2898.110 0.313986156
#> 848   2900.730 0.312527661
#> 849   2903.350 0.294415194
#> 850   2905.970 0.254097432
#> 851   2908.600 0.219330723
#> 852   2911.220 0.223423553
#> 853   2913.840 0.196984100
#> 854   2916.460 0.174708140
#> 855   2919.070 0.163559158
#> 856   2921.690 0.170455216
#> 857   2924.310 0.142659131
#> 858   2926.920 0.153726285
#> 859   2929.540 0.160633861
#> 860   2932.150 0.180038530
#> 861   2934.760 0.152256944
#> 862   2937.370 0.117543139
#> 863   2939.980 0.124469197
#> 864   2942.590 0.091153326
#> 865   2945.200 0.071721177
#> 866   2947.810 0.012372822
#> 867   2950.420 0.031486276
#> 868   2953.020 0.019010947
#> 869   2955.630 0.023196064
#> 870   2958.230 0.006570288
#> 871   2960.830 0.017708043
#> 872   2963.440 0.024689665
#> 873   2966.040 0.009902002
#> 874   2968.640 0.012750199
#> 875   2971.240 0.000000000
#> 876   2973.830 0.000000000
#> 877   2976.430 0.000000000
#> 878   2979.030 0.007068246
#> 879   2981.630 0.000000000
#> 880   2984.220 0.014752200
#> 881   2986.810 0.000000000
#> 882   2989.410 0.008093181
#> 883   2992.000 0.000000000
#> 884   2994.590 0.000000000
#> 885   2997.180 0.000000000
#> 886   2999.770 0.000000000
#> 887   3002.360 0.000000000
#> 888   3004.950 0.011969174
#> 889   3007.530 0.000000000
#> 890   3010.120 0.000000000
#> 891   3012.700 0.000000000
#> 892   3015.290 0.006680356
#> 893   3017.870 0.000000000
#> 894   3020.450 0.000000000
#> 895   3023.040 0.000000000
#> 896   3025.620 0.000000000
#> 897   3028.200 0.000000000
#> 898   3030.770 0.000000000
#> 899   3033.350 0.000000000
#> 900   3035.930 0.007462537
#> 901   3038.500 0.000000000
#> 902   3041.080 0.000000000
#> 903   3043.650 0.000000000
#> 904   3046.230 0.009133126
#> 905   3048.800 0.000000000
#> 906   3051.370 0.000000000
#> 907   3053.940 0.000000000
#> 908   3056.510 0.000000000
#> 909   3059.080 0.000000000
#> 910   3061.650 0.000000000
#> 911   3064.210 0.006455428
#> 912   3066.780 0.000000000
#> 913   3069.350 0.000000000
#> 914   3071.910 0.000000000
#> 915   3074.470 0.000000000
#> 916   3077.040 0.009101920
#> 917   3079.600 0.000000000
#> 918   3082.160 0.000000000
#> 919   3084.720 0.000000000
#> 920   3087.280 0.000000000
#> 921   3089.830 0.000000000
#> 922   3092.390 0.008295141
#> 923   3094.950 0.000000000
#> 924   3097.500 0.010098521
#> 925   3100.060 0.000000000
#> 926   3102.610 0.000000000
#> 927   3105.160 0.000000000
#> 928   3107.720 0.000000000
#> 929   3110.270 0.007065966
#> 930   3112.820 0.000000000
#> 931   3115.370 0.000000000
#> 932   3117.910 0.010168534
#> 933   3120.460 0.009974135
#> 934   3123.010 0.000000000
#> 935   3125.560 0.000000000
#> 936   3128.100 0.007075860
#> 937   3130.640 0.007951929
#> 938   3133.190 0.008129390
#> 939   3135.730 0.012843771
#> 940   3138.270 0.012998816
#> 941   3140.810 0.012451942
#> 942   3143.350 0.011194920
#> 943   3145.890 0.000000000
#> 944   3148.430 0.007154299
#> 945   3150.960 0.017178869
#> 946   3153.500 0.018981461
#> 947   3156.030 0.015271431
#> 948   3158.570 0.018554625
#> 949   3161.100 0.028257911
#> 950   3163.640 0.022407447
#> 951   3166.170 0.016187043
#> 952   3168.700 0.024746268
#> 953   3171.230 0.030171531
#> 954   3173.760 0.034271894
#> 955   3175.300 0.037342083
#> 956   3177.840 0.037393623
#> 957   3180.380 0.043096159
#> 958   3182.910 0.037684523
#> 959   3185.450 0.046011509
#> 960   3187.990 0.037684523
#> 961   3190.520 0.037684523
#> 962   3193.060 0.043235847
#> 963   3195.590 0.043235847
#> 964   3198.120 0.032133199