Identify and filter spectra

match_spec() joins two OpenSpecy objects and their metadata based on similarity. cor_spec() correlates two OpenSpecy objects, typically one with knowns and one with unknowns. ident_spec() retrieves the top match values from a correlation matrix and formats them with metadata. get_metadata() retrieves metadata from OpenSpecy objects. max_cor_named() formats the top correlation values from a correlation matrix as a named vector. filter_spec() filters an Open Specy object.

Usage

cor_spec(x, ...)

# S3 method for default
cor_spec(x, ...)

# S3 method for OpenSpecy
cor_spec(x, library, na.rm = T, ...)

match_spec(x, ...)

# S3 method for default
match_spec(x, ...)

# S3 method for OpenSpecy
match_spec(
  x,
  library,
  na.rm = T,
  top_n = NULL,
  order = NULL,
  add_library_metadata = NULL,
  add_object_metadata = NULL,
  fill = NULL,
  ...
)

ident_spec(
  cor_matrix,
  x,
  library,
  top_n = NULL,
  add_library_metadata = NULL,
  add_object_metadata = NULL,
  ...
)

get_metadata(x, ...)

# S3 method for default
get_metadata(x, ...)

# S3 method for OpenSpecy
get_metadata(x, logic, rm_empty = TRUE, ...)

max_cor_named(cor_matrix, na.rm = T)

filter_spec(x, ...)

# S3 method for default
filter_spec(x, ...)

# S3 method for OpenSpecy
filter_spec(x, logic, ...)

ai_classify(x, ...)

# S3 method for default
ai_classify(x, ...)

# S3 method for OpenSpecy
ai_classify(x, library, fill = NULL, ...)

Arguments

x

an OpenSpecy object, typically with unknowns.

library

an OpenSpecy or glmnet object representing the reference library of spectra or model to use in identification.

na.rm

logical; indicating whether missing values should be removed when calculating correlations. Default is TRUE.

top_n

integer; specifying the number of top matches to return. If NULL (default), all matches will be returned.

order

an OpenSpecy used for sorting, ideally the unprocessed one; NULL skips sorting.

add_library_metadata

name of a column in the library metadata to be joined; NULL if you don't want to join.

add_object_metadata

name of a column in the object metadata to be joined; NULL if you don't want to join.

fill

an OpenSpecy object with a single spectrum to be used to fill missing values for alignment with the AI classification.

cor_matrix

a correlation matrix for object and library, can be returned by cor_spec()

logic

a logical or numeric vector describing which spectra to keep.

rm_empty

logical; whether to remove empty columns in the metadata.

...

additional arguments passed cor().

Value

match_spec() and ident_spec() will return a data.table-class() containing correlations between spectra and the library. The table has three columns: object_id, library_id, and match_val. Each row represents a unique pairwise correlation between a spectrum in the object and a spectrum in the library. If top_n is specified, only the top top_n matches for each object spectrum will be returned. If add_library_metadata is is.character, the library metadata will be added to the output. If add_object_metadata is is.character, the object metadata will be added to the output. filter_spec() returns an OpenSpecy object. cor_spec() returns a correlation matrix. get_metadata() returns a data.table-class()

with the metadata for columns which have information.

See also

Author

Win Cowger, Zacharias Steinmetz

Examples

data("test_lib")

unknown <- read_extdata("ftir_ldpe_soil.asp") |>
  read_any() |>
  conform_spec(range = test_lib$wavenumber,
               res = spec_res(test_lib)) |>
  process_spec()
cor_spec(unknown, test_lib)
#>                                     intensity
#> 00002e4e3fac430aa1fdfea6e26f85e4  0.569407508
#> 00061dd49cbb71549cf3d530b894ba92 -0.052190336
#> 0008a60c1af45a76ffb91b9cbe1a32e4  0.207979965
#> 000902ae526db452960c5782843081b4 -0.056716505
#> 00133c1f1531821f815431f72871eae9  0.340306755
#> 0016677a99407c18717bf5d3eea9a0a2  0.115573415
#> 00185609a3205a19225e1530ca98084e -0.007595939
#> 001d3eb19976c785442a3e6fa8094bff -0.006349272
#> 0026eee747d277edf308194c81179cb0 -0.055994097
#> 0031bb13faea1e04b52ffbeca009e8ab  0.667212394
#> 003b2e57d47a4225b8ea041c946cfc0a -0.021394278
#> 0046ff947759e247f85dc63a2a25e097  0.028093717
#> 0054792a45a11ccf7158fdf2c8873125  0.086561975
#> 005c6a81975d747fe493032cac9dfaa2  0.208633314
#> 0062d71f901713f2ac7a89c72fb186d4  0.283591721
#> 00904e33ccbaa20fb68c1943d8303d82  0.008448131
#> 00931a5ccadb549d463293381c974d79 -0.019239797
#> 00b9f87c7d7c82675cb945fe4da37917 -0.084062797
#> 00f4b09b9823187caf65b781b8be87b9 -0.094082248
#> 010bfd4be6c24ced909f6de717c7b04e  0.075922217
#> 015e8a21344a60a020226cfef344c584  0.027354156
#> 016961c0d7def12f68239756d23a79a6 -0.073087780
#> 019b8f3e4839f8fda1441f7d712ff12d -0.008758977
#> 01b2fea6718855567e2e468a792b457f  0.052087746
#> 02b8194bd714469666151a7c6b90b36c  0.067350361
#> 03f1fa06cd6bc468ae5e406ca1d58dbb -0.056403149
#> 041771facd71667d6bd6c9269d3701be  0.048507385
#> 0481958e37c72c71f88d80e4dd69ac59  0.040059596
#> 058a9af32053d14e876703a60a37498f -0.017563249
#> 05c3cde45098e01cc44da3f72ada913e -0.058071417
#> 06a2ed935a9b37b653717ff4a0ea62f7  0.051112510
#> 06f7950194c7a42714cec488e378a820  0.104432379
#> 087b8dc9fe95b7cab9d037304f0d39c8 -0.007197481
#> 09b69fcfcf3c695299dd0f88743cf14a  0.068682329
#> 0b10e6cc01c8e53f2221647a9c828164  0.547532315
#> 0b8b5f10d7a9b2bd7d702cd4fdfd8a29  0.019559339
#> 0e815860f0115ac34e2b46ac91e0c096  0.021717032
#> 101b6ae86864958ddd95445d0ab01fe4  0.428785312
#> 14ee1d85d0b1ee84fab0773329c45bd2 -0.053850428
#> 1a31cfc8e25332f4771d03e8a862e217  0.025145450
#> 23069d13c090ee5235874bd632467f9c  0.023152174
#> 252b6f9c5fb83ec82246bc041cabd221 -0.016101759
#> 263479214867d17e135a0d599b1f8354  0.495372081
#> 2fa630bec93db824509be188494d79ad  0.099396099
#> 35ce7ae774d57362dacb92cacd53b01c  0.006021983
#> 3d35fd9a2766afaf4c088b615dffc350 -0.001855729
#> 4762edfa07e6b28a4b90db83594f2b3d -0.001305780
#> 4f1d701f4080b4c74acdaa04dea52dd1  0.066564239
#> 5961a72282786fe43d8704780782f459  0.063100625
#> 65792c9e80934e4692540a3e0fbfb552  0.046166444
#> 6e013532253522149a0deb04373b79a2  0.132547463
#> 7167a50e76f4a34dbb1d354babd674b0  0.038050661
#> 76b3af06ab610907cea4dbe3aa4ec67e  0.024080222
#> 889adce6238a677669b6523d74635d03  0.125236768
#> 9f91f6441608ecca35f69d298e5351bd  0.001937168
#> a692fd0b8b3b80b640ce4bbea8105570  0.044361185
#> ec69350b1b7de35fd6b8cf65048e5fc0  0.104947519

match_spec(unknown, test_lib, add_library_metadata = "sample_name",
           top_n = 1)
#>                          library_id object_id match_val  x y
#> 1: 0031bb13faea1e04b52ffbeca009e8ab intensity 0.6672124 37 1
#>                         SpectrumID   Organization SpectrumType SpectrumIdentity
#> 1: HDPE_Sarah_#18_1s_20ac_10x_25mW J. Lynch, NIST        Raman             HDPE
#>    Polymer.Category LibraryType OWNER SpectralCollectionMode Preprocessing
#> 1:      polyolefins    Polymers  <NA>                   <NA>          <NA>
#>    InstrumentUsed RRUFFID IDEAL CHEMISTRY LOCALITY SOURCE STATUS  URL
#> 1:           <NA>    <NA>            <NA>     <NA>   <NA>   <NA> <NA>
#>    MEASURED CHEMISTRY OtherInformation JCAMP-DX YFACTOR YUNITS InstrumentMode
#> 1:               <NA>             <NA>     <NA>    <NA>   <NA>           <NA>
#>    DELTAX Sponsor Color Cell   ID Comment cell_plate_id
#> 1:   <NA>    <NA>  <NA> <NA> <NA>    <NA>          <NA>
#>    Form..film..foam.pliable..hard. Brand Item Location Notes Longest.dimension
#> 1:                            <NA>  <NA> <NA>     <NA>  <NA>              <NA>
#>    Width Depth Particle.mass..mg. Tg..oC. Tm..oC. Onset.... Enthalpy..J.g.
#> 1:  <NA>  <NA>               <NA>    <NA>    <NA>      <NA>           <NA>
#>    Substance Form Colour Filename SourceDatabase Method...49 File QA/QC
#> 1:      <NA> <NA>   <NA>     <NA>           <NA>        <NA> <NA>  <NA>
#>    Natural /Synthetic Plastic/other framework Abbreviation Source ID
#> 1:               <NA>          <NA>      <NA>         <NA>      <NA>
#>    Method...57 morphology color final_polymer_assignment Citation Plate
#> 1:        <NA>       <NA>  <NA>                     <NA>     <NA>    NA
#>    Itteration SpectralResolution InstrumentAccessories Wavenumber_Range PIN_ID
#> 1:         NA               <NA>                  <NA>             <NA>   <NA>
#>    ORIENTATION ID merged database Database ID WWTP Paper/Kirstie/FLOPPE...70
#> 1:        <NA>                 NA                                         NA
#>    Database ID WWTP Paper/Kirstie/FLOPPE...71 ID (ESM1) Longest_dimension width
#> 1:                                         NA        NA                NA    NA
#>    Cluster2018/WWTP PRESSURE TEMPERATURE MaterialForm MaterialProducer
#> 1:               NA       NA          NA         <NA>             <NA>
#>    NumberofAccumulations TIME Ratio against background Truncated
#> 1:                    NA <NA>                     <NA>      <NA>
#>    ElementMultiply or ElementDivide Collection length Apodization
#> 1:                             <NA>              <NA>        <NA>
#>    Bench serial number Subtraction or Addition Baseline Correction Smooth
#> 1:                <NA>                    <NA>                <NA>   <NA>
#>    %Transmittance->Absorbance Sample gain Number of background scans
#> 1:                       <NA>          NA                         NA
#>    Background gain Sample spacing Number of scan points Resolution points
#> 1:              NA             NA                    NA                NA
#>    Number of FFT points ZPD Library ID Shortform Grating Hole size (nm)
#> 1:                   NA  NA                 <NA>      NA             NA
#>    Slit (um) Filter (%) Delay (s) DATE MOLFORM $NIST SOURCE $NIST DOC FILE
#> 1:        NA         NA        NA <NA>    <NA>         <NA>           <NA>
#>    $NIST PSD FILE APERTURE BEAMSPLITTER DETECTOR (DIA. DET. PORT IN SPHERE)
#> 1:           <NA>     <NA>         <NA>                                <NA>
#>    SPHERE DIAMETER ACQUISITION MODE SCANNER SPEED PHASE CORRECTION ZEROFILLING
#> 1:            <NA>             <NA>          <NA>             <NA>        <NA>
#>    WAVENUMBER ACCURACY LOW PASS FILTER SWITCH GAIN ON $NIST ID COADDED SCANS
#> 1:                <NA>            <NA>           <NA>     <NA>            NA
#>    PHASE RESOLUTION MW ContactInfo CASNumber MaterialQuality LaserLightUsed
#> 1:               NA NA        <NA>      <NA>            <NA>           <NA>
#>    TotalAcquisitionTime_s DataProcessingProcedure
#> 1:                   <NA>                    <NA>
#>    LevelofConfidenceinIdentification smoother baseline DENSITY new_label
#> 1:                              <NA>       NA       NA      NA      <NA>
#>                polymer_class plastic_or_not
#> 1: Polyolefins (POLYALKENES)        plastic
#>                                                   url_polymer_class
#> 1: https://www.polymerdatabase.com/polymer%20index/polyalkenes.html
#>           polymer                                                url_polymer
#> 1: POLY(ETHYLENE) https://www.polymerdatabase.com/polymers/polyethylene.html
#>                                                               url_more_info
#> 1: https://www.polymerdatabase.com/polymer%20classes/Polyolefin%20type.html
#>    snr_deriv                          file_id
#> 1:  1339.684 77f837c1640910f2879184d61ef1df48